% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{add.media}
\alias{add.media}
\title{add.media}
\usage{
add.media(n, media, pd)
}
\arguments{
\item{n}{Number of persons.}

\item{media}{List of the potential contact media for the model; output of the \code{\link{read.media.file}} function. 
Each is found in a specific microenvironment (micro).}

\item{pd}{Data frame containing internally assigned demographic and physiological parameters for each theoretical person
modeled in SHEDS.HT. Output of the \code{\link{select.people}} function}
}
\value{
pdm A data table containing the \code{pd} data frame of physiological and demographic parameters for each theoretical
person, and the \code{dur} array, which specifies the duration of exposure to each potential exposure medium for each person
in  \code{pd}.
}
\description{
For each theoretical person parameterized in the \code{pd} data frame, which is output from the \code{\link{select.people}}
function, this function generates the exposure duration for each potential exposure medium.
}
\details{
To generate the exposure duration for each theoretical person, an array q of uniform random samples is generated with
rows = \code{n} (per set, where set size is specified by the user in the \code{\link{Run}} input file) and columns =
\code{nrow(media)} (the number of potential exposure media). An array called \code{dur} is created for the number of minutes on
the activity diary in the relevant micro multiplied by the relevant probability of contact (as specified in the
\code{media} input). The array consists of a row for each person and a column for each of the exposure media.  The output is a
data table containing the \code{pd} data frame and the \code{dur} array.
}
\seealso{
\code{\link{select.people}}, \code{\link{read.media.file}}
}
\author{
Kristin Isaacs, Graham Glen
}
