% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_temporal_analyte_polluter.R
\name{add_temporal_analyte_polluter}
\alias{add_temporal_analyte_polluter}
\title{Some title}
\usage{
add_temporal_analyte_polluter(
  path_total,
  df_analyte_to_append_filepath,
  df_polluter_to_append_file_path,
  upstream_thresh = 5,
  downstream_lower_thresh = 0,
  downstream_upper_thresh = 50,
  num_cores = 1,
  permission = "no"
)
}
\arguments{
\item{path_total}{description of this parameter}

\item{df_analyte_to_append_filepath}{description of this parameter}

\item{df_polluter_to_append_file_path}{description of this parameter}

\item{upstream_thresh}{description of this parameter}

\item{downstream_lower_thresh}{description of this parameter}

\item{downstream_upper_thresh}{description of this parameter}

\item{num_cores}{description of this parameter}

\item{permission}{description of this parameter}
}
\value{
description of the return value
}
\description{
A description of this function with inline latex \eqn{2^3=8}
test for displayed equations \deqn{\frac{1}{2}\sum_{i=1}^n\log(x+y)}.
We can also use links, tables, lists and Character formattings such as bold and italic.
}
