% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{addFortyTwo}
\alias{addFortyTwo}
\title{Adds the number 42 to the specified number.}
\usage{
addFortyTwo(num)
}
\arguments{
\item{num}{The number to add 42 to or a vector of numbers.}
}
\value{
The specified number plus 42 added to it or a
numeric vector with the same length as the input and
each element being plus 42.
}
\description{
If only a single number is specifed, then the computation is done
in pure R code. If a vector of numbers with length greater
than 1 is specified, then the number 42 is added to each element
of the vector and the computation is performed in native code.
}
\details{
This is pretty self-explanatory.
}
\examples{
\dontrun{
# This is a comment
x <- addFortyTwo(5)
y <- addFortyTwo(c(12, 34, 56))
}

}
