% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addVecFortyTwoNative}
\alias{addVecFortyTwoNative}
\title{Adds the number 42 to each element of the specified numeric vector.}
\usage{
addVecFortyTwoNative(vec)
}
\arguments{
\item{vec}{A numeric vector for which each element
should be incremented by 42.}
}
\value{
A numeric vector with the same length as the input,
but each element is incremented by 42.
}
\description{
Here we could give some more information about this
extraordinary function, but we think it's pretty self-explanatory.
}
\details{
This is pretty self-explanatory.
}
\examples{
\dontrun{
// This is a comment
x <- addVecFortyTwoNative(c(11, 22, 33, 44))
}

}
