% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_cat.R
\name{age_cat}
\alias{age_cat}
\title{Catagorizes a basic bins for grouping and report generation}
\usage{
age_cat(age = 0, age_group = "adult")
}
\arguments{
\item{age}{is a patient age number of year in decimal}

\item{age_group}{is a text valuse for "adult, combo, blind or peds"}
}
\value{
returns the age bin.
}
\description{
Original function was designed for pediatric age binning this
iteration will look at all but subdivision of the pediatric
population will depend on the type of center.
}
\details{
Very basic function at this point need verification of adult bins.
age cut-off recommendations and attempt to have large enough bin
cohort size to decrease probability of individual determination.
}
\examples{
\dontrun{

> age_cat(22)
#[1] "20's"

> age_cat(44)
#[1] "40's"

> age_cat(44, age_group = "blind")
#[1] "18 to under 60"

> age_cat(44, age_group = "combo")
#[1] "40's"

}



}
\references{
\url{https:\\ <Enter git hub or other reference>}
}
\seealso{
\code{\link{age_catagory_old.R}}
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
\keyword{acs}
\keyword{age,}
