% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{allstats.variable.rank.plot}
\alias{allstats.variable.rank.plot}
\title{allstats.variable.rank.plot}
\usage{
allstats.variable.rank.plot(
  run.name = specs$run.name,
  output.variable = "abs.tot.mgkg",
  metrics = c("50\%", "95\%", "mean"),
  cohort.col = "Total",
  label_chem = F
)
}
\arguments{
\item{run.name}{default = name of last run (in current R session)}

\item{output.variable}{default = abs.tot.mgkg (mg/kg/day).}

\item{metrics}{argument is a vector of character strings with desired percentiles or metrics.}

\item{cohort.col}{argument vector of different cohorts. default = "Total". To add cohorts use a vector of concatenated character strings, such as cohort.col = c("Total", "Male", "Female")}
}
\value{
a plot of output.variable vs chemical rank
}
\description{
A ggplot scatter plot with output.variable on the y axis and chemical rank on the x axis.
User can color points by cohort (see cohort.col argument) and assign percentiles to be points (see metrics argument)
}
\details{
A SHEDS run creates an output file for each chemical. This plot pulls from the allstats.csv for each chemical in the output file.
Chemicals are ranked in ascending order of output.variable on the x axis. The y axis is log transformed and a constant of 1e-10 added to y variable for
log transformation of 0's. The default color is black for the "Total" cohort, but user can include as many cohorts as they wish as long as they are calculated in the SHEDS run.
Percentiles must be calculated in SHEDS run.the selected cohorts.
}
