% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSE__functions.R
\name{analyse.calculate.means}
\alias{analyse.calculate.means}
\title{Generic function for calculating means}
\usage{
analyse.calculate.means(fixreport_df, aggregation_column_list,
  output_column_expression, final_output_column_expression, spss,
  dvColumnName, prefixLabel = "", debug = FALSE)
}
\arguments{
\item{fixreport_df}{Fixation report}

\item{aggregation_column_list}{List of columns to group by}

\item{output_column_expression}{Output column expression}

\item{final_output_column_expression}{Final output column expression}

\item{spss}{Should the function output for SPSS?}

\item{dvColumnName}{Column name of the dependent variable}

\item{prefixLabel}{Prefix label}

\item{debug}{Should debug information be provided?}
}
\value{
A data.table ready for SPSS analyses, which is also saved to disk as a text file.
}
\description{
Generic function for calculating means
}
\examples{
# THIS IS A UTILITY FUNCTION THAT YOU WOULD NOT NORMALLY USE YOURSELF
}
