% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{black}
\alias{black}
\title{Black's measles seroprevalence data.}
\format{
A data frame with 42 rows and 3 variables:
\describe{
  \item{age}{ age-bracket (in years)}
  \item{mid}{mid-point of age-bracket (in years)}
  \item{n}{number of tests}
  \item{pos}{number seropositive}
  \item{neg}{number seronegative}
  \item{f}{seroprevalence}
}
}
\source{
Black (1959) Measles antibodies in the population of New Haven, Connecticut. Journal of Immunology 83:74-83
}
\usage{
black
}
\description{
Seroprevalence-by-age-bracket for measles in prevaccination New Haven as studied by Black (1959).
}
\keyword{datasets}
