% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data frame which consists the value and the weight for each object.}

\item{W}{The knapsack size.}

\item{parallel}{Gives the opportunity to use parallel programming,if parallel=TRUE.}
}
\value{
The functoin returns a list containing two named objects:
\itemize{
  \item "value": maximum knapsack value;
  \item "elements": a vector containing the indexes of the objects (rows of data.frame) used to obtain the final result.
}
}
\description{
this function is guaranteed to give a correct answer in all situations.
The function enumerates all different combinations by using a binary representation of the numbers
1 to 2n and include all elements of that is equal to 1 in the binary representation.
}
