% --- Source file: data.Rd ---
\name{BSR.obj}
\alias{BSR.obj}

\title{
    Return object from \code{\link{BayesSigRef}}
}
\description{
    An example return object from \code{\link{BayesSigRef}}.
}
\details{ 
    See \code{\link{BayesSigRef}} for details.
}
\examples{

    data(BSR.obj, package="BayesSigRefitting")
    names(BSR.obj)

}
\keyword{data}
\seealso{ \code{\link{BayesSigRef}} }


