% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnett}
\alias{burnett}
\title{Burnett's Parasitoid-Host data.}
\format{
A data frame with 22 rows and 7 variables:
\describe{
  \item{Generation}{ }
  \item{NumberofHostsExposed}{ }
  \item{NumberofHostsParasitized}{ }
  \item{NumberofHostsUnparasitized}{ }
  \item{NumberofParasiteEggsLaid}{ }
  \item{NumberofParasitesSearching}{ }
  \item{PercentageofHostsParasitized}{ }
}
}
\source{
Burnett, T. A. (1958) Model of host-parasite interaction Proceedings of the 10th International Congress, Entomology, 1958, 2, 679-686
}
\usage{
burnett
}
\description{
Data is of 22 generations of greenhouse white flies (\emph{Trialeurodes vaporariorum}) and its parasitoid, \emph{Encarsia formosa}.
Column names are self explanatory.
}
\keyword{datasets}
