% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccs}
\alias{ccs}
\title{UK measles CCS data.}
\format{
A data frame with 954 rows and 14 variables:
\describe{
\item{fade3}{Average duration of fadeout (of at least 3 weeks of length)}
  \item{ext}{Fraction of time when measles was absent}
  \item{size}{Median population size}
  \item{fade}{Average duration of fadeouts (of a week or longer)}
  \item{se3}{Standard error fade3}
  \item{se}{Standard error of fade}
  \item{n3}{The number of fadeouts (of at least 3 weeks of length)} 
  \item{n}{The number of fadeout of a week or longer}
  \item{names}{City/town name}
}
}
\source{
Bjornstad and Grenfell (2008) Hazards, spatial transmission and 
timing of outbreaks in epidemic metapopulations. 
Environmental and Ecological Statistics 15: 265-277. <doi:10.1007/s10651-007-0059-3>
}
\usage{
ccs
}
\description{
The fraction of weeks measles was absent from each of the 954 cities
and towns of England and Wales between 1944 and 1965.
}
\keyword{datasets}
