% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center2circle.R
\name{center2circle}
\alias{center2circle}
\title{Center to circle function list (center2circle)}
\usage{
center2circle(lon = NA, lat = NA, name = NA, geo_id = NA,
  rad = ..., unit = ...)
}
\arguments{
\item{lon}{is the longitude of the center}

\item{lat}{is the latitude of the center}

\item{name}{is a basic unique id that can be name or number used
for group id}

\item{geo_id}{is the name use to look up the lon, lat via geo code}

\item{rad}{is the radius of the circle}

\item{unit}{unit of measure of the radius}
}
\value{
returns a list with necessary attributes and the data frame of circle points
}
\description{
Given a data from or an data set of at least 1 to extract the lon
lat and add a ID.  Then the placement of 1-360 degrees around the
radius is calculated using the geosphere package.  The information
is then placed into a list and then returned.
}
\examples{
/donotrun{
#Do not run
using wv_trauma.csv

circle.60 <- ctr2cir(data.set = wv_trauma, rad = 60, unit = "mi")

# Running with lapply

#XXX ADD!!!!

#End not run

}
}
\references{
\url{https://stackoverflow.com/questions/34183049
/plot-circle-with-a-certain-radius-around-point-on-a-map-in-ggplot2}

add geosphere reference
}
\seealso{
\code{\link{ctr2cir}}

\code{ctr2cir.R}
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
