% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fugacity.R
\name{chem.fug}
\alias{chem.fug}
\title{chem.fug}
\usage{
chem.fug(n.per, cprops, x)
}
\arguments{
\item{n.per}{The total number of simulated persons in this model run specified in the \code{\link{Run}} file.}

\item{cprops}{The chemical properties required for SHEDS-HT (output of \code{chem.props} function). The default data was
prepared from publicly available databases using a custom program (not part of SHEDS-HT).
The default file contains about 17 numerical inputs per chemical, but most are not used.
The required properties are molecular weight (\code{MW}), vapor pressure (\code{VP.Pa}),
solubility (\code{water.sol.mg.l}), octanol-water partition coefficient (\code{log.Kow}),
air decay rate (\code{half.air.hr}), decay rate on surfaces (\code{half.sediment.hr}),
and permeability coefficient (\code{Kp}).}

\item{x}{From the output of the \code{\link{read.fug.inputs}} function.}
}
\value{
samples A data set with the chemical specific parameters for each combination of chemical and simulated person.
For each chemical, the chemical specific parameters assigned to a given person are randomly sampled from distributions
on those parameters. These distributions are created from point estimates to reflect real-world uncertainty and variability.
}
\description{
Creates distributions of chemical specific parameters for each chemical of interest, in order to reflect real-world
variability and uncertainty. These distributions are then sampled to create chemical specific parameters associated
with the exposure of each simulated person.
}
\details{
This function obtains chemical-specific properties from the chem.data data set.That data set contains
point value for each variable, which this function defines distributions around, with the exception of molecular
weight. The constructed distributions reflect both uncertainty and variability (for example, vapor pressure will
vary with temperature, humidity, and air pressure or altitude, and may depend on the product formulation).
For each variable, a random sample is generated for each simulated person, and the set of variables becomes the
output data set.
Input surface loading variables are in units of ug/cm2 or ug/cm2/day and are converted by the function to meters,
to avoid the need for conversion factors in later equations. These conversions are done after the random sampling,
as otherwise the correct conversions depend on the distributional form (for example, par2 is changed for the normal,
but not for the lognormal).
}
\seealso{
\code{\link{Fugacity}}, \code{\link{Run}}, \code{\link{run}}, \code{\link{Chem_props}}, \code{\link{get.fug.concs}}
}
\author{
Kristin Isaacs, Graham Glen
}
\keyword{SHEDS}
