% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{chem.scenarios}
\alias{chem.scenarios}
\title{chem.scenarios}
\usage{
chem.scenarios(all)
}
\arguments{
\item{all}{This is the master list of all chemicals, and all exposure scenarios specific to each chemical, to be evaluated
in the current model run. This data set is compiled internally according to the user's specifications on the
\code{\link{Run_85687}} file. The user may create multiple scenarios files for special purposes, for example, for selected
chemical classes, or selected exposure pathways. A model run consists of two nested loops: the outer loop over chemicals and
the inner loop over the scenarios specific to that chemical.}
}
\value{
\item{chem.scen }{A data set consisting of all chemicals of interest, and the categories and scenarios for that chemical
with the potential for exposure.  The relationship between chemicals and scenarios is one to many. It is possible for one
chemical to have two or more "dermal" scenarios, provided each is in a different "category".  However, each combination of
chemical, category, and scenario must be unique. The last variable on chem.scen is "count", which is the number of rows of data
on all.scenarios devoted to this combination.}
}
\description{
This function summarizes the all.scenarios data set by condensing each chemical-scenario combination into one row. The
number of rows of data for this combination on all.scenarios is recorded here.
}
\seealso{
\code{\link{Run}}, \code{\link{Source_chem_foods}}, \code{\link{Source_chem_prods}}, \code{\link{Source_scen_food}},  \code{\link{Source_scen_prods}}
}
\author{
Kristin Isaacs, Graham Glen
}
