% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholera}
\alias{cholera}
\title{Dacca cholera death data.}
\format{
A data frame with 600 rows and 4 variables:
\describe{
  \item{Year}{Year}
  \item{Month}{Month of the year}
\item{Dacca}{Monthly cholera deaths}
  \item{Population}{Population size of district}
}
}
\source{
King, A.A., Ionides, E.L., Pascual, M. and Bouma, M. J. (2008) 
Inapparent infections and cholera dynamics. Nature, 454:877-880. \url{http://doi.org/10.1038/nature07084}
}
\usage{
cholera
}
\description{
Monthly deaths from  cholera in Dacca, East Bengal between 1891 and 1940.
}
\keyword{datasets}
