\name{cobindr-internal}

\alias{compare.samples}
\alias{cpg.content}
\alias{determine.cores.option}
\alias{detrending}

\alias{find.pairs_intern}
\alias{find.hits}
\alias{gc.content}
\alias{cpg.gc.content}
\alias{generate.background}
\alias{get.positions}
\alias{global.permutation}

\alias{input.pfm.similarity}
\alias{input.pwm}
\alias{query.motifDb}

\alias{kl.dist}
\alias{lcm.vector}
\alias{local.permutation}
\alias{make.character}

\alias{normalize.matrix.colSum}
\alias{query.motifDb}
\alias{parallelize}
\alias{pfm.similarity}
\alias{pfm2pwm}

\alias{read.background.fasta}
\alias{read.pfm}
\alias{read.sequences}
\alias{read.transfac.pfm}
\alias{removeSpecialCharacters}
\alias{rtfbs.intern}

\alias{uniqueFilename}
\alias{wind.cpg.content}
\alias{wind.gc.content}
\alias{wind.cpg.gc.content}
\alias{writeDetrending_intern}

\title{Internal cobindr Functions}
\description{
 Internal cobindr functions
}

\details{
These are not to be called by the user.
}
\keyword{ internal }