% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_groups.R
\name{count_groups}
\alias{count_groups}
\title{Count Observations per Groups}
\usage{
count_groups(x, ...)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{...}{Columns in \code{x}.}
}
\value{
A \code{\link{data.frame}} with the counts per groups.
}
\description{
I tend to forget the syntax that works with
\code{\link[stats:aggregate]{stats::aggregate}}.
}
\examples{
count_groups(mtcars, "am", "gear")
RUnit::checkEquals(dplyr::count(mtcars, am, gear),
                   count_groups(mtcars, "am", "gear"), checkNames = FALSE)
}
\seealso{
Other statistics: 
\code{\link{column_sums}()},
\code{\link{relative_difference}()},
\code{\link{round_half_away_from_zero}()},
\code{\link{weighted_variance}()}
}
\concept{statistics}
