% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr2cir.R
\name{ctr2cir}
\alias{ctr2cir}
\title{Center to circle function basic (ctr2cir)}
\usage{
ctr2cir(data.set = ..., rad = 45, unit = "mi")
}
\arguments{
\item{data.set}{Data set that contains at a minimum a lon & lat columns}

\item{rad}{radius in miles, meters or kilometers}

\item{unit}{unit of measure of the radius}
}
\value{
returns a data frame ID = N in data frame lon and lat of the points of the circle
}
\description{
Given a data from or an data set of at least 1 to extract the lon
lat and add a ID.  Then the placement of 1-360 degrees around the
radius is calculated using the geosphere package.
}
\details{
The ctr2cir function differs from the center2circle function in
that its more of a minimalist function that just returns a data
frame and group ID of the circle points from the lon, lat, radius
(rad) and unit of measurement for the radius.  This function is
basically the Stack Overflow code.
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
