% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dalziel}
\alias{dalziel}
\title{Measles incidence across 40 US cities}
\format{
A data frame with 44,720 rows and 10 variables:
\describe{
  \item{biweek}{biweek of the year}
  \item{cases}{incidence}
  \item{year}{year}
  \item{loc}{city name}
  \item{pop}{population size}
  \item{rec}{susceptible recruits}
  \item{country}{country}
  \item{lon}{city longitude} 
  \item{lat}{city latitude} 
  \item{decimalYear}{time counter}
}
}
\source{
Dalziel et al. 2016. Persistent chaos of measles epidemics in the prevaccination United States caused by a small change in seasonal transmission patterns. PLoS Computational Biology 2016: e1004655.  <doi:10.1371/journal.pcbi.1004655>
}
\usage{
dalziel
}
\description{
A dataset of Measles incidence across 40 US cities with relevant demographic data
}
\keyword{datasets}
