% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\docType{data}
\name{data.ann}
\alias{data.ann}
\title{Gene expression of a two condition time series experiment}
\format{A data frame with 12762 rows and 36 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75417}
}
\usage{
data.ann
}
\description{
A dataset containing the gene expression of a time series experiment
with 6 time points (0,2,6,12,18,24), 2 conditions and 3 biological replicates
}
\keyword{datasets}
