
\name{data}
\alias{data}



%\docType{package}
\title{
  Data for examples
}
\description{
 Data for examples.
}
\details{ A data frame used in the examples. 
}
\examples{

 data(data, package="Immunotherapy.Design")

 # Display some of the data
 data[1:5, ]
}
\keyword{data}
%\seealso{ \code{\link{subhmm}} }



