% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{dir.ingested}
\alias{dir.ingested}
\title{dir.ingested}
\usage{
dir.ingested(sd, cd)
}
\arguments{
\item{sd}{The chemical-scenario data specific to relevant combinations of chemical and scenario. Generated internally.}

\item{cd}{The list of scenario-specific information for the chemicals being evaluated. Generated internally.}
}
\value{
dir.ingest For each person, the calculated quantity of a given chemical incidentally ingested during or
immediately after use of products such as toothpaste. Does not include exposure via food and drinking water.
}
\description{
Models the ingestion exposure scenario for each theoretical person.
}
\details{
This scenario is for accidental ingestion during product usage. Typical examples are toothpaste, mouthwash, 
lipstick or chap stick, and similar products used on the face or mouth.
The function produces a \code{prevalence} value, which reflects the fraction of the population who use this scenario at all.
It also produces a \code{frequency} value, which is the mean number  of times per year this scenario occurs among that
fraction of the population specified by prevalence.
Since SHEDS operates on the basis of one random day, the frequency is  divided by 365 and then passed to the
\code{\link{p.round}} (probabilistic rounding) function, which rounds either up or down to the nearest integer. Very common
events may happen more than once in a day.
The function also produces a \code{mass} variable, which refers to the  mass of the product in grams in a typical usage event.
The \code{composition} is the percentage of that mass that is the chemical in question.
The \code{ingested} variable represents the percentage of the mass applied that becomes ingested. Since these products are
not intended to be swallowed, this should typically be quite small (under 5\%). 
The final output is the total incidental ingested exposure for each chemical-individual combination in micrograms, which
is the product of the above variables multiplied by a factor of 1E6.
}
\author{
Kristin Isaacs, Graham Glen
}
