% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{dir.inhal.vap}
\alias{dir.inhal.vap}
\title{dir.inhal.vap}
\usage{
dir.inhal.vap(sd, cd, cprops, cb, io)
}
\arguments{
\item{sd}{The chemical-scenario data specific to relevant combinations of chemical and scenario. Generated internally.}

\item{cd}{The list of scenario-specific information for the chemicals being evaluated. Generated internally.}

\item{cprops}{The chemical properties required for SHEDS-HT. The default file (the \code{\link{Chem_props file}} read in by
the \code{\link{read.chem.props}} function and modified before input into the current function) was prepared from publicly
available databases using a custom program (not part of SHEDS-HT).  The default file contains 7 numerical inputs per
chemical, and the required properties are molecular weight (\code{MW}), vapor pressure (\code{VP.Pa}), solubility
(\code{water.sol.mg.l}), octanol-water partition coefficient (\code{log.Kow}), air decay rate (\code{half.air.hr}),
decay rate on surfaces (\code{half.sediment.hr}), and permeability coefficient (\code{Kp}).}

\item{cb}{A copy of the \code{base} data set output from the \code{\link{make.cbase}} function, with columns added
for exposure variables.}

\item{io}{A binary variable indicating whether the volume of the aerosol is used to approximate the affected volume.}
}
\description{
Models the inhalation exposure from the vapors of volatile chemicals for each theoretical person.
}
\details{
This scenario considers inhalation exposure from vapors (not aerosols). For example, painting will result in the
inhalation of vapor, but it does not involve aerosols (unless it is spray paint).  
For this scenario, the vapor pressure and the molecular weight are relevant variables for determining exposure. These
variables are included in the input to the \code{cprops} argument, which is drawn internally from the
\code{\link{Chem_props}} file.
The function produces a \code{prevalence} value, which reflects the fraction of the population who use this scenario at
all. It also produces a \code{frequency} value, which is the mean number of times per year this scenario occurs among that
fraction of the population specified by \code{prevalence}. 
Since SHEDS operates on the basis of one random day, the \code{frequency} is  divided by 365 and then passed to the
\code{\link{p.round}} (probabilistic rounding) function, which rounds either up or down to the nearest integer. Very
common events may happen more than once in a day.
The function also produces a \code{mass} variable, which refers to the  mass of the product in grams in a typical usage
event. The \code{composition} is the percentage of that mass that is the chemical in question. The \code{evap} variable is
an effective evaporated mass, calculated using the \code{mass}, \code{composition} (converted from percent to a fraction),
the vapor pressure as a surrogate for partial pressure, and \code{duration} of product use.  The \code{duration} term is
made unitless by dividing by 5 (minutes), which is an assumed time constant.
The effective air concentation \code{airconc} is calculated as \code{evap}/\code{volume}. The value for \code{airconce}
is capped by \code{maxconc}, which represents the point at which evaporation ceases.  For chemicals used for a short
duration, or with low vapor presssure, \code{maxconc} might not be reached before usage stops.
Once \code{airconc} is established, the function also calculates the inhaled dose, in units of micrograms per day.
The dose equals the product of \code{exposure} (g/m3), basal ventilation rate, \code{bvr} (m3/day), the METS factor of 1.75
(typically people inhale air at an average of 1.75 times the basal rate to support common daily activities), and a
conversion factor of 1E6 from grams to micrograms.
}
\seealso{
\code{\link{run}}, \code{\link{p.round}}, \code{\link{read.chem.props}}, \code{\link{Chem_props}}
}
\author{
Kristin Isaacs, Graham Glen
}
