% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_cat.R
\name{disp_cat}
\alias{disp_cat}
\title{Dispensation category function}
\usage{
disp_cat(disp = "Floor")
}
\arguments{
\item{disp}{is the dispensation of the patient}
}
\value{
Text to simplify patient dispensation.
}
\description{
Given a disp column will break up into basic groups that stay in
the hospital, discharged, mortality or other.
}
\details{
The goal of this function is to take a dispensation and reduce the
number of factors to basic groups. Initially, started with four
groups (i.e. Stay in Hospital, Morgue, transfer or home).
}
\examples{
\dontrun{
  ### to call using purr::### map_chr(peds.hvmc$ISS, ISS.cat)
}

}
\references{

}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
\keyword{dispensation,}
\keyword{simplification}
