% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{dmvn}
\alias{dmvn}
\title{Probability density function for multivariate normal distribution}
\usage{
dmvn(x, mean, sigma, log = TRUE)
}
\arguments{
\item{x}{Observation vector}

\item{mean}{Multivariate normal distribution mean vector}

\item{sigma}{Multivariate normal distribution variance/covariance matrix}

\item{log}{The probability value returns as log-value. Defaults to TRUE.}
}
\value{
The probability value of 'x'
}
\description{
This function computes the probability density function for a multivariate normal distribution
}
\examples{
x <- c(1.5, 4.5)
mean <- c(1,4)
sigma <- matrix(c(1,0.7,0.7,1), byrow=TRUE, ncol = 2)
dmvn(x, mean , sigma)
}
\keyword{probabilityDensityFunction}
