% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSex.R
\name{estimateSex}
\alias{estimateSex}
\title{estimateSex Function}
\usage{
estimateSex(beta.value = NULL, detecP = NULL, beta.intervals.X = seq(0, 1,
  0.1), beta.intervals.Y = seq(0, 1, 0.1), p.intervals.X = c(-18, -5, -2,
  0), p.intervals.Y = c(-18, -5, -2, 0), return_with_reference = FALSE)
}
\arguments{
\item{detecP}{detection p-value matrix}

\item{beta.intervals.X}{Intervals of beta-value of X chromosome probes within each of which the proportion is calculated. Default is 'seq(0,1,0.1)', which is 0 to 0.1, 0.1 to 0.2, ..., 0.9 to 1.0.}

\item{beta.intervals.Y}{Intervals of beta-value of Y chromosome probes within each of which the proportion is calculated. Default is the same as that of 'beta.intervals.X'}

\item{p.intervals.X}{Intervals of detection p-value of X chromosome probes within each of which the proportion is calculated. Default is -Inf to 1e-18m 1e-18 to 1e-5, 1e-5 to 1e-2, 1e-2 to 0.}

\item{p.intervals.Y}{Intervals of detection p-value of Y chromosome probes within each of which the proportion is calculated. Default is the same as that of 'p.intervals.X'}

\item{return_with_reference}{Logical value to indicate whether or not to return the data derived from references data (Default to FALSE).}

\item{beta-value}{beta-value matrix}
}
\description{
This function estimates the sex of DNA methylation microarray data frmo the distribution of beta-value and detection p-value of methylation sites from X chromosome and Y chromosome.
}
\keyword{estimateSex}
