% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
returns greater common divisor of two value
}
\description{
Given 2 values (x and y), the function choose the greatest (for example x) one and replace it with modular of the other (x mod y) until remainer is equal to zero.
}
\examples{
euclidean(100, 1000) 
}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{wikipedia page}
}
