% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euthamia}
\alias{euthamia}
\title{Euthamia graminifolia rust data.}
\format{
A data frame with 360 rows and 8 variables:
\describe{
  \item{block}{the block}
  \item{row}{row}
  \item{plot}{plot within block}
  \item{xloc}{x coordinates}
  \item{yloc}{y coordinate}
  \item{comp}{plot composition}
  \item{water}{treatment: dry or wet}
  \item{score}{the rust score}
}
}
\usage{
euthamia
}
\description{
Data on a fungal pathogen of the aster Euthamia graminifolia collected by Jennifer Keslow.
}
\keyword{datasets}
