% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{eval.factors}
\alias{eval.factors}
\title{eval.factors}
\usage{
eval.factors(r, q, ef)
}
\arguments{
\item{r}{Vector of row numbers, equivalent to the number of theoretical people in the model sample. This input is created
internally by the \code{\link{add.factors}} function.}

\item{q}{User-specified list of desired quantiles to be included in the output.}

\item{ef}{Distributional parameters for the exposure factors; an output of the \code{\link{exp.factors}} function and an
input argument to the \code{\link{add.factors}} function}
}
\value{
z A data frame specifying the form of the distribution and relevant parameters for each combination of exposure
factor and individual person. The parameters include:
\describe{
\item{\code{form}}{The form of the distribution for each exposure factor (i.e., point, triangle).}
\item{\code{par1-par4}}{The parameters associated with the distributional form specified in form.}
\item{\code{lower.trun}}{the lower limit of values to be included in the distribution.}
\item{\code{upper.trun}}{The upper limit of values to be included in the distribution.}
\item{\code{resamp}}{Logical field where: yes=resample, no=stack at truncation bounds.}
\item{\code{q}}{The quantiles of the distribution corresponding to those specified by the user in the \code{q} input.}
\item{\code{p}}{The probabilities associated with the quantiles.}
\item{\code{v}}{The values associated with the quantiles.}
}
}
\description{
Assigns distributions for each relevant exposure factor for each theoretical persons to be modeled in SHEDS.HT.
}
\seealso{
\code{\link{add.factors}}, \code{\link{exp.factors}}
}
\author{
Krisin Isaacs, Graham Glen
}
