% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{featureSelectionOneConditionHMM}
\alias{featureSelectionOneConditionHMM}
\title{One condition feature selection hidden Markov model}
\usage{
featureSelectionOneConditionHMM(data, times, reps, LOG = TRUE, t.size = 1,
  iters = 50, USE.REP = T, func = median, factor = 10, prior = FALSE,
  initial.vector = NULL, transition.matrix = NULL)
}
\arguments{
\item{data}{numeric matrix - Expression data (Log2 data)}

\item{times}{numeric vector - All the time points when the data was measured}

\item{reps}{numeric - How many biological replicates do we have per time point?}

\item{LOG}{logical - Do you wanna use the data in its Log2 transformation or set them in its natural range? Defaults to TRUE.}

\item{t.size}{numeric - What percentage of data will be used to get estimate the parameters? Range - (0,1]. Defaults to .}

\item{iters}{numeric - How many iterations will be used to estimate the model parameters? Defaults to 50.}

\item{USE.REP}{logical - Want to use all the replicates as input? Defaults to TRUE.}

\item{func}{function - If USE.REP == TRUE, then how are they going to be summarized? (mean, median, customFunction). Defaults to median.}

\item{factor}{numeric - Factor to multiply the data. Usually used recommended to be used with the LOG = T. Defaults to 10.}

\item{prior}{logical - Do you want to use a prior for the model parameters? Defaults to FALSE.}

\item{initial.vector}{vector - stochastic vector with the model initial probability vector prior. Defaults to NULL.}

\item{transition.matrix}{matrix - stochastic matrix with the model transition probability matrix prior. Defaults to NULL.}
}
\value{
final.results a list consisting of: 
\item{data}{matrix - relevant and filtered data} 
\item{names}{vector - relevant features names} 
\item{model}{list - HMM used to select the most relevant data} 
\item{scoreReplicates}{vector - The score of biological replicates per feature} 
\item{scoreChanges}{vector - the score of the feature given the number of times the it traverse a 'change state'} 
\item{scoreDelta}{vector - the score given the proportion of the changes in time per feature} 
\item{scoreMaxDelta}{ - vector - the score of the feature given its maximum change in time } 
\item{ranking}{vector - the rank of the feature given the replicate, changes and delta scores} 
\item{decodedStates}{matrix - each row has the decoded states of each change per feature }
}
\description{
This function computes a feature selection filtering via a hidden Markov model
}
\details{
This function uses a hidden Markov model with an observation vector 
that can be modeled with a multivariate normal distribution for one condition
to select the best features of the observation matrix.
}
\examples{


data("ratTimeSeries")

# All the necessary variables to run the script
times <- c(0,2,6,12,18,24)  # study time-points vector
conds <- c("CTRL","IKAROS") # Condition vector
reps <- 3                   # number of biological replicates
\dontrun{
# It has some random sampling inside, so a set.seed is recommended
# Case: Biological replicates summarized by median
set.seed(100)
summ.results <- featureSelectionOneConditionHMM(data.ann[,19:36],      # Gene expression data matrix
                                                times,         # time-point vector
                                               reps,          # Number of replicates
                                               USE.REP  = F,  # Do not use all replicate data
                                               func = median  # Summarize the data
)

# Case: Biological replicates not summarized
set.seed(100) 
mv.results <- featureSelectionHMM(data.ann[,19:36],      # Gene expression data matrix
                                 times,         # time-point vector
                                 reps           # Number of replicates
)  

plotMultivariate(mv.results, times, conds[2], reps, ".", "Fat.Rat.multi.all.pdf")
plotUnivariate(summ.results, times, conds[2], reps, ".", "Fat.Rat.uni.res.pdf")
}

}
\seealso{
\code{\link{featureSelectionHMM}} for multiple condition feature selection.
}
\keyword{FeatureSelection}
