% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferrari}
\alias{ferrari}
\title{Ferrari et al. 2005 outbreak data.}
\format{
A data frame with 15 rows and 7 variables:
\describe{
  \item{Eboladeaths00}{Number of deaths from ebola during the 2000 Uganda outbreak}
  \item{Ebolacases00}{Number of cases of ebola during the 2000 Uganda outbreak}
  \item{Ebolacases95}{Number of cases of ebola during the 1995 DRC outbreak}
  \item{FMDfarms}{Number of farms infected with FMD during the 2000-01 UK outbreak}
  \item{HogCholera}{Number of cases of swine fever in pigs in the 1997-98 outbreak in the Netherlands}
  \item{SarsHk}{Number of cases of SARS in Hong Kong during the 2003 outbreak}
  \item{SarsSing}{Number of cases of SARS in Singapore during the 2003 outbreak}
}
}
\source{
Ferrari et al. (2005) Estimation and inference of R-0 of an infectious pathogen by a removal method. Mathematical Biosciences 198: 14-26. \url{http://doi.org/10.1016/j.mbs.2005.08.002}
}
\usage{
ferrari
}
\description{
The incidence aggregated by serial interval of a number of outbreaks studied by Ferrari et al. 2005.
}
\keyword{datasets}
