% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filipendula}
\alias{filipendula}
\title{Filipendula rust data.}
\format{
A data frame with 162 rows and 4 variables:
\describe{
  \item{y94}{infection status in 1994}
  \item{y95}{infection status in 1995}
  \item{X}{X coordinate}
  \item{Y}{Y coordinate}
}
}
\source{
Smith et al. 2003. Epidemiological patterns at multiple spatial scales: an 11-year study of a Triphragmium ulmariae -- Filipendula ulmaria metapopulation. Journal of Ecology, 91(5), pp.890-903.  \url{http://doi.org/10.1046/j.1365-2745.2003.00811.x}
}
\usage{
filipendula
}
\description{
Rust infection status of 162 populations of \emph{Filipendula ulmaria}
in a Swedish Island archipelago
}
\keyword{datasets}
