% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fiv}
\alias{fiv}
\title{FIV infection in cats.}
\format{
A data frame with 238 rows and 18 variables:
\describe{
  \item{Id}{Individual identifier}
  \item{CD4}{CD4 cell count}
  \item{CD8B}{CD8B cell count}
  \item{CD25}{CD25 cell count}
  \item{FAS_L}{FAS ligand}
  \item{FAS}{FAS}
  \item{IFNg}{Interferon gamma}
  \item{IL_10}{Interleukin 10}
  \item{IL_12}{Interleukin 12}
  \item{IL_4}{Interleukin 4}
  \item{lymphocyte}{lymphocyte count}
  \item{neutrophils}{neutrophil count}
  \item{TNF_a}{Tumor necrosis factor}
  \item{provirus}{provirus count}
  \item{viremia}{viremia}
  \item{Day}{day}
  \item{No}{unique identifier}
  \item{Treatment}{Experimental treatment}
}
}
\source{
Roy et al. 2009. Multivariate statistical analyses demonstrate unique host immune responses to single and dual lentiviral infection. PLoS one 4, e7359. \url{http://doi.org/10.1371/journal.pone.0007359}
}
\usage{
fiv
}
\description{
Immunological measures on cats infected with different strains of FIV
}
\keyword{datasets}
