% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationreport.R
\docType{data}
\name{fixationreport}
\alias{fixationreport}
\title{Example fixation report dataset}
\format{
A data.table object.
}
\usage{
data(fixationreport)
}
\description{
Fixation report data from a visual search experiment.
}
\examples{
data(fixationreport)

}
\references{
TBA.
}
\keyword{datasets}
