% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu}
\alias{flu}
\title{Boarding school influenza data.}
\format{
A data frame with 14 rows and 2 variables:
\describe{
  \item{day}{day since beginning of outbreak}
  \item{cases}{number of sick children}
}
}
\source{
Anonymous (1978) EPIDEMIOLOGY: Influenza in a boarding school. British Medical Journal, 4 March 1978 p.587.
}
\usage{
flu
}
\description{
The daily number of children confined to bed in a boarding school in North England during an outbreak in 1978 of the reemerging A/H1N1 strain.
The school had 763 boys of which 512 boys were confined to bed sometime during the outbreak.
}
\keyword{datasets}
