% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{food.migration}
\alias{food.migration}
\title{food.migration}
\usage{
food.migration(cdata, sdata, cb, ftype)
}
\arguments{
\item{cdata}{The list of scenario-specific information for the chemicals being evaluated. Generated internally.}

\item{sdata}{The chemical-scenario data specific to relevant combinations of chemical and scenario. Generated internally.}

\item{cb}{A copy of the \code{base} data set output from the \code{\link{make.cbase}} function, with columns added for
exposure variables.}

\item{ftype}{Food consumption database which stores data on consumption in grams per day of each food type for each person
being modeled. Generated internally from the  \code{\link{Diet_diaries}} data set.}
}
\value{
dietary The calculated quantity of chemical exposure from food residue and migration of chemicals into food from 
packaging and other contact materials in grams per person per day.
}
\description{
Calculates chemical exposure from migration of chemicals from packaging or other contact materials into food (which is
then consumed). This is added to the \code{dietary} output calculated by the \code{\link{food.residue}} function to
establish a total exposure from the dietary pathway.
}
\details{
If migration data is stored in the \code{cdata} argument, this is added to the total exposure calculated in the
\code{\link{food.residue}} function.
}
\seealso{
\code{\link{Diet_diaries}}, \code{\link{run}}, \code{\link{p.round}}, \code{\link{make.cbase}}, \code{\link{food.residue}}
}
\author{
Kristin Isaacs, Graham Glen
}
