% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{food.residue}
\alias{food.residue}
\title{food.residue}
\usage{
food.residue(cdata, cb, ftype)
}
\arguments{
\item{cdata}{The list of scenario-specific information for the chemicals being evaluated. Generated internally.}

\item{cb}{Output of the \code{\link{make.cbase}} function.}

\item{ftype}{Food consumption database which stores data on consumption in grams per day of each food type for each
person being modeled. Generated internally from the  \code{\link{Diet_diaries}} data set.}
}
\value{
dietary The calculated quantity of chemical exposure from food residue in grams per person per day.
}
\description{
Models exposure to chemicals from consumption of food containing a known chemical residue for each theoretical person.
}
\details{
In this function, the variable \code{foods} is defined as a list of the names of the food groups,
which are stored in both the \code{ftype} and \code{cb} arguments. The FOR loop picks the name of each food group as
a string, and converts it to a variable name. For example, the food group "FV" may have corresponding variables
\code{residue.FV}, \code{zeros.FV} (number of nondetects), and \code{nonzeros.FV} (number of detects) on the \code{cb}
data set. If these variables are not present, no  exposure results.  The variables may be present, but an individual
person may still receive zero exposure because  not all samples of that food are contaminated, as indicated by the
\code{zeros.FV} value.  The \code{nonzeros.FV} value is used to  determine the likelihood of contamination, with the
residue variable determining the amount found (when it is nonzero).
The dietary exposure is the product of the \code{consumption} as reported in the \code{ftypes} input (in grams) and
the \code{residue.FV} (in micrograms of chemical per gram of food), summed over all food groups. Three other vectors are
returned (corresponding to dermal exposure, inhalation exposure, and inhalation dose), but these are currently set to zero
for the dietary pathway. In principle, eating food could result in dermal exposure (for finger foods), or inhalation (as
foods may have noticeable odors), but such exposures are not large enough to be of concern at present.
}
\seealso{
\code{\link{Diet_diaries}}, \code{\link{run}}, \code{\link{p.round}}, \code{\link{generate.person.vars}}, \code{\link{food.migration}}
}
\author{
Kristin Isaacs, Graham Glen
}
