% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritools-package.R
\docType{package}
\name{fritools-package}
\alias{fritools-package}
\title{Utilities for the Forest Research Institute of the State
Baden-Wuerttemberg}
\description{
Miscellaneous utilities, tools and helper functions.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_fritools", package = "fritools")}.
}
\keyword{package}
