\name{gcdisp}
\alias{gcdisp}
\docType{data}
\title{average GC displacement from [1]}
\description{average GC displacement from [1] in degrees of visual angle (0.3 mm per degree of visual angle [2])}
\usage{data( gcdisp )}
\format{
Displacement is a function of eccentricity of the GC soma position.
\describe{
\item{\code{ecc}}{eccentricity of the GC soma position}
\item{\code{displ}}{displacement}
  }
}

\keyword{datasets}

\references{
[1] N. Drasdo, C. L. Millican, C. R. Katholi, and C. A. Curcio. The length of Henle
fibers in the human retina and a model of ganglion receptive field density in the
visual field. Vision Research, 47:2901-2911, 2007.

[2] B. A. Wandell. Foundations of vision. Sinauer Associates, Sunderland,
Massachusets, 1995.
}

\author{Ivan Marin-Franch}