% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{gen.factor.tables}
\alias{gen.factor.tables}
\title{gen.factor.tables}
\usage{
gen.factor.tables(ef = exp.factors)
}
\arguments{
\item{ef}{A data set created internally using the \code{\link{run}} function and the \code{\link{read.exp.factors}} 
function to import the user specified \code{\link{Exp_factors}} input file. The data set contains the distributional
parameters for the exposure factors. All of these variables may have age or gender-dependent distributions, although
in the absence of data, many are assigned a single distribution from which all persons are sampled.}
}
\value{
exp.gen Output consists of non-media specific exposure factors as a data table. Depending on the user input,
these generated exposure factors may be gender and/or season specific.  Each gender-season combination is assigned
a row number pointing to the appropriate distribution on the output dataset from the \code{\link{read.exp.factors}} function.
Thus, \code{exp.gen} consists of 8 rows per variable (2 genders x 4 seasons), regardless of the number of different
distributions used for that variable.
In addition, if the function runs successfully, the following message is printed: "General Factor Tables completed"
}
\description{
Constructs tables of non-media specific exposure factors for each relevant combination of age, gender, and season. The 
input is from the \code{\link{Exp_actors}} file (specified on the \code{\link{Run}} file) after being read through the 
\code{\link{read.exp.factors}} function.
}
\seealso{
\code{\link{Exp_factors}}, \code{\link{Run}}, \code{\link{run}}, \code{\link{read.exp.factors}}
}
\author{
Kristin Isaacs, Graham Glen
}
