% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READING__generate.R
\name{generate.reading.trial}
\alias{generate.reading.trial}
\title{Generate Reading Trial Image and Interest Area File}
\usage{
generate.reading.trial(
  preTargetText = "Life was a fly that faded and death a",
  targetText = " drone",
  postTargetText = " that stung. ",
  width = 1024,
  height = 768,
  fontSize = 14,
  letterWidth = 8,
  font = "courier new",
  fontColor = "black",
  backgroundColor = "white",
  trialLabel = "sentence1",
  iaFolder = getwd(),
  imgFolder = getwd(),
  startX = 50,
  startY = 768/2,
  highlightEachBox = FALSE
)
}
\arguments{
\item{preTargetText}{The sentence text before the target word.}

\item{targetText}{The sentence text comprising the target word only.}

\item{postTargetText}{The sentence text that follows the target word.}

\item{width}{The width of the display in pixels (default 1024).}

\item{height}{The height of the display in pixels (default 768)}

\item{fontSize}{The font size of the text (default 14)}

\item{letterWidth}{The letter width. This is used to space the letters out if they overlap.}

\item{font}{The font to use (default is Courier New)}

\item{fontColor}{The colour of the font to use (default is black)}

\item{backgroundColor}{The background colour of the display (default is white)}

\item{trialLabel}{The label of your trial. This will be used to name your image and associated interest area file.}

\item{iaFolder}{The folder to save your interest area files into.}

\item{imgFolder}{The folder to save your images into.}

\item{startX}{The start position on the x axis for your sentence.}

\item{startY}{The start position on the y axis for your sentence.}

\item{highlightEachBox}{When set to TRUE, this puts a randomly-coloured box behind each letter. 
Used for making sure that letterWidth is set to something sensible. Default is FALSE.}
}
\value{
String saying 'DONE' when completed. Also saves an interest area file and an image.
}
\description{
Generate Reading Trial Image and Interest Area File
}
