% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfstats.R
\name{getage}
\alias{getage}
\title{Calculates age}
\usage{
getage(dob, date)
}
\arguments{
\item{dob}{date(s) of birth}

\item{date}{date(s) for which to calculate age}
}
\value{
\code{getage} returns the age from the date of birth and a certain date
}
\description{
Computes ages at specific dates
}
\examples{
getage("1977-01-31", "2014-01-30")
}
