
\name{getBenchmarkGenes}
\alias{getBenchmarkGenes}

\title{Benchmark Genes}
\description{
  Get a vector of gene names given by Bailey et al (2018)                    
}
\usage{
getBenchmarkGenes()
}
%\arguments{}

\details{
  This is the list of driver gene names given by Bailey et al (2018). 
}
\value{
  A character vector of gene names.
}
%\references{
%  Choo-Wosoba H., Albert P.S., Zhu B.
%  A Hidden Markov Modeling Approach for Identifying Tumor Subclones in Next-Generation Sequencing Studies.
%}

%\author{Ho-Hsiang Wu and Bin Zhu <bin.zhu@nih.gov>}
\examples{
  getBenchmarkGenes()
}
\keyword{ mutation } 
%\seealso{\code{\link{MutScotPatientHeter}}}
