\name{ getnv }
\alias{ getnv }
\title{get normative values and location map}
\description{get normative values and stimuli location map to be used with visualFields}
\usage{getnv()}
\details{
gets normative values and stimuli location map to be used with visualFields
}

\value{
returns the names of the variables containing the normative values and the stimuli location map
}

\seealso{ \code{\link{setnv}} }

\examples{
  getnv()
}

\author{Ivan Marin-Franch}