% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gpars.R
\docType{data}
\name{gpars}
\alias{gpars}
\title{List of graphical parameters}
\format{
See section \code{Structure of graphical parameters} in \code{\link{vfplot}}
}
\usage{
gpars
}
\description{
It contains a list of normative values, including pointwise and
smoothed SUNY-IU normative reference values for 24-2 static automated perimetry
(\code{sunyiu_24d2_pw} and \code{sunyiu_24d2}) obtained with the dataset
\code{\link{vfctrSunyiu24d2}}
}
\keyword{dataset}
