% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Title greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data frame which consists the value and the weight for each object.}

\item{W}{The knapsack size}
}
\value{
The functoin returns a list containing two named objects:
\itemize{
  \item "value": maximum knapsack value;
  \item "elements": a vector containing the indexes of the objects (rows of data.frame) used to obtain the final result.
}
}
\description{
This function uses Greedy heuristic.
It does not return an exact result but it will reduce the computational complexity.
}
