% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gypsymoth}
\alias{gypsymoth}
\title{Defoliated by gypsy moth each in northeast US 1975-2002.}
\format{
A list with two matrices each with 1086 rows:
\describe{
  \item{xy}{A matrix with two columns representing UTM coordinates}
 \item{defoliation}{A matrix with 28 columns representing pixel-wise defoliation between 1976 and 2002}
}
}
\source{
Bjornstad, O. N., Robinet, C., & Liebhold, A. M. (2010). Geographic variation in North American gypsy moth cycles: subharmonics, generalist predators, and spatial coupling. Ecology, 91(1), 106-118. \url{http://doi.org/10.1890/08-1246.1}
}
\usage{
gypsymoth
}
\description{
A list containing the fraction of forest defoliated by the gypsy moth
in 20km x 20km pixels across northeast US in each year between 1975 and 2002.
}
\keyword{datasets}
