
\name{hseg_data}
\alias{hseg_data}
\alias{lr}
\alias{logor}


%\docType{package}
\title{
  Data for examples
}
\description{
 Data for examples.
}
\details{ The object contains \code{logR} and \code{logOR} values
  for the \code{\link{hsegHMM_T}} and \code{\link{hsegHMM_N}} examples.
}
\examples{

 data(hseg_data, package="hsegHMM")

 # Display some of the data
 lr[1:10]
 logor[1:10]
}
\keyword{data}
\seealso{\code{\link{hsegHMM_T}}, \code{\link{hsegHMM_N}} }



