% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_cat.R
\name{icd_cat}
\alias{icd_cat}
\title{Assigns an injury group based upon ICD 10 code name}
\usage{
icd_cat(ICD = ...)
}
\arguments{
\item{ICD}{is the location or one code to look up.}
}
\value{
returns the overall group the injury is assigned to
    (i.e. MVC, FARI, fall, NFPI (non firearm penetrating injury),
    blunt, animal, poison, etc.)
}
\description{
Given one or a vector of ICD-10 codes for the primary injury use
the overall categories to group.
}
\details{
This is a basic minimal viable product due to difficult obtaining
all the ICD 10 codes at the time of initial authorship. The list
was compiled by looking at the unique diagnosis.  Goal is to later
obtain all the trauma injury codes and develop a lookup table.
First version was written using unique entries and assignment.
Need to add non-firearm penetrating injury (NFPI), blunt, animal
and others as needed.
}
\examples{
\dontrun{

# Just one

test_icd <- ("all-terrain")

icd_cat(ICD = test_icd)

# Map across and entire data frame

purr::map_chr(trauma_data$ISS, icd_cat)

}


}
\references{
\url{https:\\ <Enter git hub or other reference>}
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
\keyword{ICD,}
\keyword{grouping}
