% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icelandflu}
\alias{icelandflu}
\title{Monthly incidence of influenza-like illness in Iceland between 1980 and 2009.}
\format{
A data frame with 360 rows and 3 variables:
\describe{
  \item{month}{the month}
  \item{year}{the year}
  \item{ili}{ILI incidence}
}
}
\source{
Bjornstad ON, Viboud C. Timing and periodicity of influenza epidemics. Proceedings of the National Academy of Sciences. 2016 Nov 15;113(46):12899-901. \url{http://doi.org/10.1073/pnas.1616052113}
}
\usage{
icelandflu
}
\description{
A dataset containing the monthly ILI incidence 
in Iceland between 1980 and 2009.
}
\keyword{datasets}
