% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{indir.exposure}
\alias{indir.exposure}
\title{indir.exposure}
\usage{
indir.exposure(sd, cb, concs, chem.data)
}
\arguments{
\item{sd}{The chemical-scenario data specific to relevant combinations of chemical and scenario. Generated internally.}

\item{cb}{A copy of the \code{base} data set output from the \code{\link{make.cbase}} function, with columns added for
exposure variables.}

\item{concs}{The concentration of the chemical (in air and/or on surfaces) being released into the environment. Outpu of
the \code{\link{get.fug.concs}} function.}

\item{chem.data}{The list of scenario-specific information for the chemicals being evaluated. Generated internally.}
}
\value{
indir Indirect exposure to chemicals in the home in ug per person per day.
}
\description{
Models the indirect exposure to chemicals in the home for each theoretical person.
}
\details{
Indirect exposure happens after a product is no longer being used or applied, due to chemical lingering on various
surfaces or in the air. People who come along later may receive dermal or inhalation exposure from residual chemical in the
environment.
SHEDS.HT currently has two indirect exposure scenarios. One which applies to a one-time chemical treatment applied to a
house, and another which applies to continual releases from articles. Both scenarios consist of two parts: the first
determines the appropriate air and surface concentrations. That code is in the \code{\link{Fugacity}} module.  The second
part is the exposure calculation by the current function. Both types of indirect exposure scenarios call this function.
The surface and air concentrations from the \code{\link{Fugacity}} module are premised on the product use actually occurring.
Hence, \code{\link{indir.exposure}} starts by multiplying those concentrations by the \code{prevalence} (which is either 0
or 1, evaluated separately for each person).
For air, the exposure is the average daily concentration, which is the event concentration multiplied by the fraction of
the day spent in that event. The inhaled dose is the product of the exposure, the basal ventilation rate, and the PAI factor
(multiplier for the basal rate). A factor of 1E+06 converts the result from grams per day to micrograms per day.
Dermal exposure results from skin contact with surfaces. The surface concentration (ug/cm2) is multiplied by the
fraction available for transfer (\code{avail.f}, unitless), the transfer coefficient (\code{dermal.tc}, in cm2/hr), and the
contact \code{duration} (hr/day).  The result is the amount of chemical transferred onto the skin (ug/day).
}
\seealso{
\code{\link{Fugacity}}, \code{\link{get.fug.concs}}, \code{\link{make.cbase}}, \code{\link{run}}
}
\author{
Kristin Isaacs, Graham Glen
}
