% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{integrandpc}
\alias{integrandpc}
\title{Auxiliary function used by llik.pc}
\usage{
integrandpc(a, up, foi)
}
\arguments{
\item{a}{a vector with the ages}

\item{up}{a vector with upper age-bracket cut-offs}

\item{foi}{a vector with FoI}
}
\value{
A vector with FoIs matched to data
}
\description{
Auxiliary function used by llik.pc
}
\seealso{
{llik.pc}
}
