% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p53sDebug.R, R/p53sf.R
\name{is.between}
\alias{is.between}
\title{Locate candidate p53 responsive elements (full and half) on a DNA sequence (working function to detect orphans)}
\usage{
is.between(x, a, b)

is.between(x, a, b)
}
\arguments{
\item{seq}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}

\item{seq}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}
}
\value{
A dataframe containing the responsive elements located on the input sequence.

A dataframe containing the responsive elements located on the input sequence.
}
\description{
Locate candidate p53 responsive elements (full and half) on a DNA sequence (working function to detect orphans)

Locate candidate p53 responsive elements (full and half) on a DNA sequence
}

