% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_of_length_zero.R
\name{is_of_length_zero}
\alias{is_of_length_zero}
\title{Is an Object of Length Zero?}
\usage{
is_of_length_zero(x, class = NULL)
}
\arguments{
\item{x}{The object.}

\item{class}{An optional character vector of length 1 giving the class. See
\emph{examples}.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Some expressions evaluate to \code{\link{integer}(0)} or the like.
}
\examples{
x <- ""; length(x); is_of_length_zero(x)
x <- grep(" ", "")
print(x)
is_of_length_zero(x)
is_of_length_zero(x, "character")
is_of_length_zero(x, "numeric")
is_of_length_zero(x, "integer")
}
\seealso{
Other logical helpers: 
\code{\link{get_run_r_tests}()},
\code{\link{is_batch}()},
\code{\link{is_cran}()},
\code{\link{is_false}()},
\code{\link{is_force}()},
\code{\link{is_installed}()},
\code{\link{is_not_false}()},
\code{\link{is_null_or_true}()},
\code{\link{is_r_cmd_check}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_running_on_fvafrcu_machines}()},
\code{\link{is_running_on_gitlab_com}()},
\code{\link{is_success}()},
\code{\link{is_version_sufficient}()},
\code{\link{is_windows}()}
}
\concept{logical helpers}
