% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss_cat.R
\name{iss_cat}
\alias{iss_cat}
\title{Injury severity score categorization function}
\usage{
iss_cat(iss_score = 0)
}
\arguments{
\item{iss_score}{}
}
\value{
category of the ISS score as per Bolorunduro et al.
}
\description{
Give an injury severity score it determines category
}
\details{
The goal of this function is to describe the patients Injury
Severity Score (ISS).  Bolorunduro et al (2011) provided for basic
categories of: Mild, Moderate, Severe and Profound.
}
\examples{
\dontrun{

# Single Call

iss_cat(8)

# Call using purr::

map_chr(peds.hvmc$ISS, ISS.cat)

}

}
\references{
Bolorunduro OB et al (2011) J Surgical Res
    2011 Mar 166 40
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}
}
\keyword{injury}
\keyword{score}
