% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ixmp.Platform.R
\docType{class}
\name{ixmp.Platform}
\alias{ixmp.Platform}
\title{Class ixmp.Platform}
\description{
The class 'ixmp.Platform' is the central access point to
and accessing dataobjects (timeseries and reference data)
and scenarios (structured model input data and results).
}
\section{Fields}{

\describe{
\item{\code{.jobj}}{"jobjRef"}
}}

\section{Methods}{

\describe{
\item{\code{close_db()}}{close the database connection of the platform instance
this is important when working with local database files ('HSQLDB') }

\item{\code{initialize(dbprops = NULL, dbtype = NULL)}}{launch the Platform}

\item{\code{open_db()}}{(re-)open the database connection of the platform instance,
e.g., to continue working after using 'close_db()}

\item{\code{Scenario(model, scen, version = NULL, scheme = NULL, annotation = NULL)}}{initialize a new Scenario (structured model input data and solution)
or get an existing Scenario from the IXMP database}

\item{\code{scenario_list(default = TRUE, model = NULL, scenario = NULL)}}{Get a list of all dataobjects (timeseries)
and scenarios initialized in the IX Modeling Platform}

\item{\code{units()}}{units function}
}}

