% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\description{
The package creates a new class called "Linreg".It consists in an alternative implemantation of the linear regression algorithm.
}
\section{Fields}{

\describe{
\item{\code{X}}{:matrix containing all the data.}

\item{\code{Y}}{:vector containing response variable.}

\item{\code{reg_coe}}{vector: Estimation of the regression coefficients.}

\item{\code{fit_val}}{vector: Estimation of the Y values.}

\item{\code{resid_e}}{vector:Estimation of the error variable.}

\item{\code{n}}{numeric :Number of data.}

\item{\code{p}}{numeric Number of variables.}

\item{\code{df}}{numeric:Degrees of freedom.}

\item{\code{resid_var}}{numeric:Estimates of the variance of the error variable.}

\item{\code{var_beta}}{matrix:Estimates the variability of the beta coefficients.}

\item{\code{t_val}}{vector:T-values for significance of coefficients.}

\item{\code{formula:Linear}}{regression formula .}

\item{\code{data:}}{Recieved data from user.}

\item{\code{dname:Data}}{name.}
}}


