% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litter}
\alias{litter}
\title{Bordetella bronchiseptica in rabbit kittens.}
\format{
A data frame with 494 rows and 8 variables:
\describe{
  \item{Facility}{breeding facility}
  \item{sick}{infection status}
  \item{Date}{date sampled}
  \item{Animal.code}{animal identifier}
  \item{msick}{dams infection status}
  \item{Litter}{litter identifier}
  \item{CFU}{bacterial count}
  \item{Description}{unique litter identifier}
}
}
\source{
Long et al (2010) Identifying the Age Cohort Responsible for Transmission in a Natural Outbreak of Bordetella bronchiseptica. PLoS Pathogens 6(12): e1001224. \url{http://doi:10.1371/journal.ppat.1001224}
}
\usage{
litter
}
\description{
Data on Bordetella bronchiseptica in rabbit kittens in a breeding facility.
}
\keyword{datasets}
