% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{llik.cb}
\alias{llik.cb}
\title{Negative log-likelihood function for the chain-binomial model}
\usage{
llik.cb(S0, beta, I)
}
\arguments{
\item{S0}{a scalar with value for S0}

\item{beta}{a scalar with value for beta}

\item{I}{a vector incidence aggregated at serial interval}
}
\value{
the negative log-likelihood for the model
}
\description{
Negative log-likelihood function for the chain-binomial model
}
\examples{
twoweek=rep(1:15, each=2)
niamey_cases1=sapply(split(niamey$cases_1[1:30], twoweek), sum)
llik.cb(S0=6500, beta=23, I=niamey_cases1)
}
