% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfloaders.R
\name{loadoctopus}
\alias{loadoctopus}
\title{Import data from Haag-Streit Eyesuite}
\usage{
loadoctopus(file, type = "pwg", repeated = mean, dateFormat = "\%d.\%m.\%Y")
}
\arguments{
\item{file}{name of the csv file exported by the eyesuite software}

\item{type}{type of patient. It can be `\code{ctr}` (for control or healthy
subject-eye) or `\code{pwg}` (for patient with glaucoma) or other}

\item{repeated}{function to apply if there are repeated values in a particular location}

\item{dateFormat}{format to be used for date. Its default value is \%d.\%m.\%Y}
}
\value{
A list with data frames and functions to extract tables.

These functions are: create_locmap(vf_id), get_sensitivities(vf_id), and get_defects(vf_id).
  
 These functions take the vf_id (which identifies different types of fields) as a parameter. This key can be found in the vf_types data frame of the returned list.
}
\description{
Parses csv-files exported from the Eyesuite Software,
is used to store and analyze data from Octopus perimeters.
}
\details{
The parser also extracts information on visual field patterns and normal values. The list that is returned contains
data frames which are structured with keys so that redundancy is minimized (similar to a relational database). Use
the functions included in the list to extract meaningful tables.
}
\seealso{
Detailed examples: \url{https://rpubs.com/huchzi/645357}

Other vfloaders: 
\code{\link{loadhfaxml}()}
}
\concept{vfloaders}
