% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadvfEyesuite.R
\name{loadvfEyesuite}
\alias{loadvfEyesuite}
\title{Import visual field data from Haag-Streit Eyesuite}
\usage{
loadvfEyesuite(filename, date_format = "\%d.\%m.\%Y")
}
\arguments{
\item{filename}{the filename of the csv-file}

\item{date_format}{the order of the dates used in the specific locale}
}
\value{
a vf object
}
\description{
\code{loadvfEyesuite} imports visual field data from the Eyesuite software by Haag-Streit. These data are converted into a vf object.
}
