% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-locmaps.R
\docType{data}
\name{locmaps}
\alias{locmaps}
\title{Location maps}
\format{
See section \code{Structure of location maps} in \code{\link{setlocmap}}
}
\usage{
locmaps
}
\description{
List of common and some custom location maps, including the
24-2, 10-2, 30-2, 60-4, etc used the the HFA and Octopus, the 24-2 used
by the Matrix (FDT), and others used in Swanson's and Wall's labs
}
\keyword{dataset}
