% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4494}
\alias{m4494}
\title{Measles in England and Wales 1944-1994.}
\format{
A list with four items:
\describe{
\item{measles}{A matrix with 354 rows and 2661 columns. Rows 
represents community and columns represent week.}
\item{ps}{A matrix with 354 rows and 41 columns. Rows 
represents community and columns represent population size 
for each of the years in the data set.}
\item{longlat}{A matrix with 354 rows and two columns. Rows 
represents community and columns longitude and latitude.}
\item{year}{A vector of length 354 that represents time for each week as yearly decimals.} 
}
}
\source{
Grenfell, B.T., Bjornstad, O.N., & Kappey, J. 2001. Travelling waves and spatial hierarchies in measles epidemics. Nature 414: 716-723. \url{http://dx.doi.org/10.1038/414716a}

Lau, M.S.Y., Becker, A.D, Korevaar, H.M., Caudron, Q., Shaw, D.J., Metcalf, C.J.E., Bjornstad, O.N. and Grenfell, B.T. 2020. A competing-risks model explains hierarchical spatial coupling of measles epidemics en route to national elimination. Nature Ecology & Evolution. \url{https://doi.org/10.1038/s41559-020-1186-6}
}
\usage{
m4494
}
\description{
The weekly reported cases of measles 
in each of the 354 cities and villages
between 1944 and 1994.
}
\keyword{datasets}
