% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magono}
\alias{magono}
\title{Massachusetts gonorrhea data.}
\format{
A data frame with 422 rows and 4 variables:
\describe{
\item{number}{Weekly case reports}
  \item{year}{Year}
  \item{week}{Week of the year}
  \item{time}{Time in fractions of year}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
magono
}
\description{
Weekly cases of gonorrhea in Massachusetts between 2006 and 2015.
}
\keyword{datasets}
