% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny2.r
\docType{data}
\name{may.app}
\alias{may.app}
\title{Launch a shiny-app simulating May's Parasitoid-host Model model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
may.app
}
\description{
Launch a shiny-app simulating May's Parasitoid-host Model model
}
\details{
Launch app for details
}
\examples{
\dontrun{may.app}
}
\keyword{datasets}
