\name{MCMCEnsembleSampler-package}
\alias{MCMCEnsembleSampler-package}
\alias{MCMCEnsembleSampler}
\docType{package}
\title{
\packageTitle{MCMCEnsembleSampler}
}
\description{
\packageDescription{MCMCEnsembleSampler}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MCMCEnsembleSampler}
\packageIndices{MCMCEnsembleSampler}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MCMCEnsembleSampler}

Maintainer: \packageMaintainer{MCMCEnsembleSampler}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
