% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messagereport.R
\docType{data}
\name{messagereport}
\alias{messagereport}
\title{Example message report dataset}
\format{A data.table object.}
\usage{
data(messagereport)
}
\description{
Message report data from a visual search experiment.
}
\examples{
data(messagereport)

}
\references{
TBA.
}
\keyword{datasets}
