% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{model_config}
\alias{model_config}
\title{Auxiliary functions configuring input and output for GAMS}
\usage{
model_config(model, case)
}
\arguments{
\item{model}{model name}

\item{case}{case}
}
\description{
Auxiliary functions configuring input and output for GAMS
}
