\name{mtest.gv}
\alias{mtest.gv}
\title{
    Tests if a 'gv' object has a model.
}
\description{
    Tests if the semi-variogram in the 'gv' object has a model fitted.
}
\usage{
    mtest.gv(gv)
}
\arguments{
  \item{gv}{
    An object of class 'gv'.
}
}
\details{
    This function is usefull to test if a model for the semi-variogram in 'gv'
    is already built.
}
\value{
    logical
}
\author{
    Pedro Tarroso <ptarroso@cibio.up.pt>
}
\note{
    It is not exported.
}
