
\name{MutSigCV.out}
\alias{MutSigCV.out}

%\docType{package}
\title{
  Example of the \code{MutSigCV.out} data object
}
\description{
  Example of the \code{MutSigCV.out} data object
}
\details{ 
 The object is a data frame named \code{MutSigCV.out}.
  Currently, the only columns of this data that are used are: \code{gene}, \code{X}, \code{x}, and \code{q}.
}
\examples{
  # Look at the first five rows 
  data(MutSigCV.out, package="MUTSCOT")
  MutSigCV.out[1:5, ]
}
\keyword{data}
%\seealso{ \code{\link{subhmm}} }



