% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sunshiny.R
\docType{class}
\name{My_shiny-class}
\alias{My_shiny-class}
\alias{My_shiny}
\title{Title my_sunshine}
\description{
This class uses another class in order to have access in an API.
It is also create a Shiny application for air pollution.
}
\section{Fields}{

\describe{
\item{\code{server_components}}{list. It contains ui and server for the Shiny application.}
}}


