% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypara.R
\docType{data}
\name{nbspat.app}
\alias{nbspat.app}
\title{Launch a shiny-app simulating the spatially-extended Nicholson-Bailey parasitoid model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
nbspat.app
}
\description{
Launch a shiny-app simulating the spatially-extended Nicholson-Bailey parasitoid model
}
\details{
Launch app for details
}
\examples{
\dontrun{nbspat.app}
}
\keyword{datasets}
