% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{networkSir}
\alias{networkSir}
\alias{netSir}
\title{Function to simulate an epidemic on a network}
\usage{
networkSir(CM, tau, gamma)
}
\arguments{
\item{CM}{a contact matrix}

\item{tau}{the transmission probability}

\item{gamma}{the recovery probability}
}
\value{
An object of class netSir with infectious status for each node through time
}
\description{
Function to simulate a stochastic (discrete time) Reed-Frost SIR model on a social network
}
\examples{
cm1=barabasiAlbert(N=200,K=2)
sim1=networkSir(cm1,.3,0.1)
summary(sim1)
\dontrun{plot(sim1)}
}
