% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{niamey}
\alias{niamey}
\title{Weekly measles incidence from 2003-04 in Niamey, Niger.}
\format{
A data frame with 31 rows and 13 variables:
\describe{
  \item{absweek}{week since beginning of outbreak}
  \item{week}{week of the year}
  \item{tot_cases}{weekly incidence for the whole city}
  \item{tot_mort}{weekly deaths for the whole city}
  \item{lethality}{weekly case fatality rate}
  \item{tot_attack}{weekly attack rates for the whole city}
  \item{cases_1}{weekly incidence for district 1}
  \item{attack_1}{weekly attack rates for district 1}
  \item{cases_2}{weekly incidence for district 2}
  \item{attack_2}{weekly attack rates for district 2}
  \item{cases_3}{weekly incidence for district 3}
  \item{attack_3}{weekly attack rates for district 3}
  \item{cum_cases}{weekly cumulative incidence for the whole city}
}
}
\source{
Grais et al (2008) Time is of the essence: exploring a measles outbreak response vaccination in Niamey, Niger. Journal of the Royal Society Interface 5: 67-74. <doi:10.1098/rsif.2007.1038>
}
\usage{
niamey
}
\description{
A dataset containing the weekly incidence of measles in
Niamey, Niger during the 2003-04 outbreak
}
\keyword{datasets}
