% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{niamey_daily}
\alias{niamey_daily}
\title{Day of appearance of each measles case from 2003-04 outbreak in Niamey, Niger.}
\format{
A data frame with 10,937 rows and 1 variables:
\describe{
  \item{day}{the day of appearance of each case since day of outbreak}
}
}
\source{
Grais et al. (2008) Time is of the essence: exploring a measles outbreak response vaccination in Niamey, Niger. Journal of the Royal Society Interface 5: 67-74. \url{http://doi.org/10.1098/rsif.2007.1038}
}
\usage{
niamey_daily
}
\description{
A dataset containing the day of appearance of each measles case  in
Niamey, Niger during the 2003-04 outbreak.
}
\keyword{datasets}
