% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{nicholsonbailey.app}
\alias{nicholsonbailey.app}
\title{Launch a shiny-app simulating the Nicholson-Bailey model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
nicholsonbailey.app
}
\description{
Launch a shiny-app simulating the Nicholson-Bailey model
}
\details{
Launch app for details
}
\examples{
\dontrun{nicholsonbailey.app}
}
\keyword{datasets}
