% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.exclusions.fix_durations}
\alias{organise.exclusions.fix_durations}
\title{Exclude very brief and very long fixations.}
\usage{
organise.exclusions.fix_durations(fixreport_df, min = 60, max = 1200)
}
\arguments{
\item{fixreport_df}{Fixation report.}

\item{min}{Minimum duration of fixations.}

\item{max}{Maximum duration of fixations.}
}
\value{
Fixation report with outliers removed. Also saves information on how many fixations were removed.
}
\description{
Exclude very brief and very long fixations.
}
\examples{
data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
                               message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
                               message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_ANSWER")

# NOW MARK UP FIXATION CONTINGENCIES
foo<-organise.message.fix_contingencies(foo, list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))

# RANDOM TRIAL TO CHECK THINGS OUT
organise.checks.random_trial(foo)

# FIX CONTINGENCIES
organise.contingencies.descriptives(foo)

# REMOVE MISSING EVENTS - HERE, TRIALS WHICH LACKED A RESPOSNE
foo <- organise.message.removals(fixreport_df=foo, 
                                 required_message_list=list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))
# REMOVES NO TRIALS = EXCELLENT!

organise.contingencies.descriptives(foo) # THIS SHOWS WE HAVE NO UNCLASSIFIED FIXATIONS, GOOD!

# BY-TRIAL BEHAVIOURAL DATA
organise.behavioural.base(fixreport_df = foo, list('PPT_ID', 'TRIALTYPE_TEXT', 'PREVALENCE'), 
                          response_period_start="SYNCTIME", 
                          prefix_label="BYTRIAL")
                          
# NEXT REMOVE PRE-SYNCTIME FIXATIONS
fooy <- foo[foo$FIXATION_CONTINGENCY!='PRE_LEADIN__LEADIN',]
fooy <- fooy[fooy$FIXATION_CONTINGENCY!='LEADIN',]
fooy <- organise.exclusions.fix_durations(fixreport_df=fooy)
}
