% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.message.descriptives}
\alias{organise.message.descriptives}
\title{Descriptive statistics for messages in message report.}
\usage{
organise.message.descriptives(message_df)
}
\arguments{
\item{message_df}{Message report.}
}
\value{
Descriptive information relating to messages in the trials sent to the console for easy viewing.
}
\description{
Descriptive statistics for messages in message report.
}
\examples{
data(messagereport)
organise.message.descriptives(messagereport)
}
