% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.message.replace_spaces}
\alias{organise.message.replace_spaces}
\title{Replace spaces in message report message with underscores.}
\usage{
organise.message.replace_spaces(message_df)
}
\arguments{
\item{message_df}{A message report.}
}
\value{
An updated message report with spaces between words replaced with underscores.
}
\description{
Replace spaces in message report message with underscores.
}
\examples{
data(messagereport)
messagereport <- organise.message.replace_spaces(messagereport)
}
