% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.message.return_specific}
\alias{organise.message.return_specific}
\title{Return trials where a specific message is found.}
\usage{
organise.message.return_specific(message_df, fixreport_df, message,
  show_working = FALSE)
}
\arguments{
\item{message_df}{Message report.}

\item{fixreport_df}{Fixation report.}

\item{message}{The message you want to search for.}

\item{show_working}{Should eyeTrackR show more detail when calculating the output?}
}
\value{
Data.table of marked up fixation report.
}
\description{
Return trials where a specific message is found.
}
\examples{
# HERE, 'SYNCTIME' STARTS A TRIAL
data(messagereport)
data(fixationreport)
organise.message.return_specific(messagereport, fixationreport, 'SYNCTIME')
}
