% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READING__functions.R
\name{organise.reading.calculateRemovals}
\alias{organise.reading.calculateRemovals}
\title{Generic function for calculating outliers in a reading study.}
\usage{
organise.reading.calculateRemovals(
  ia_df,
  source_column_name,
  output_column_name,
  filterBy = "",
  sdLimit = 2.5,
  aggregation_column_list
)
}
\arguments{
\item{ia_df}{Interest area data.frame}

\item{source_column_name}{The source column name for the outlier calculation.}

\item{output_column_name}{The output column name.}

\item{filterBy}{Which columns to filter by. This can be formatted as a 'WHERE' statement in data.table language.}

\item{sdLimit}{The SD limit - how many times the SD is the upper and lower limit? (Default is 2.5)}

\item{aggregation_column_list}{The columns to group by - can usualy be the factors in your study, or in some cases, the participants themselves.}
}
\value{
An updated interest area data.table with additional columns marked up. There will be a 'KEEP' column where 1 says it's safe to keep and 0 says to remove.
}
\description{
Generic function for calculating outliers in a reading study.
}
