% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READING__functions.R
\name{organise.reading.Skipping}
\alias{organise.reading.Skipping}
\title{Reading study - determine skipping outliers.}
\usage{
organise.reading.Skipping(ia_df, sdLimit = 2.5, aggregation_column_list)
}
\arguments{
\item{ia_df}{Interest area data.frame}

\item{sdLimit}{The SD limit - how many times the SD is the upper and lower limit? (Default is 2.5)}

\item{aggregation_column_list}{The columns to group by - can usualy be the factors in your study, or in some cases, the participants themselves.}
}
\value{
An updated interest area data.table with additional columns marked up. There will be a 'KEEP' column where 1 says it's safe to keep and 0 says to remove.
}
\description{
Reading study - determine skipping outliers.
}
