% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyorv.r
\docType{data}
\name{orv.app}
\alias{orv.app}
\title{Launch a shiny-app to study outbreak-response vaccination campaigns}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
orv.app
}
\description{
Launch a shiny-app to study outbreak-response vaccination campaigns
}
\details{
Launch app for details
}
\examples{
\dontrun{orv.app}
}
\keyword{datasets}
