% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p53maxJaspar.R
\name{p53maxJaspar}
\alias{p53maxJaspar}
\title{Return the maximum scores using jaspar p53 PWMs (pwm.1 derived from MA0106.1, built on SELEX data, pwm.2 derived from MA0106.2, built on ChIPseq data). The score ranges from 0 to 1}
\usage{
p53maxJaspar(seq)
}
\arguments{
\item{seq}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}
}
\value{
A vector containing the two maximum scores given by the two jaspar PWMs.
}
\description{
Return the maximum scores using jaspar p53 PWMs (pwm.1 derived from MA0106.1, built on SELEX data, pwm.2 derived from MA0106.2, built on ChIPseq data). The score ranges from 0 to 1
}

