% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p53plot.R
\name{p53plot}
\alias{p53plot}
\title{Plot the position p53 responsive elements on a DNA sequence (the mark is placed at the beginning of the responsive element)}
\usage{
p53plot(p53.table, start.at, stop.at)
}
\arguments{
\item{p53.table}{A dataframe containing the responsive elements, as the one returned from the p53.track function.}

\item{start.at}{A numeric value indicating the start coordinate of the plot}

\item{stop.at}{A numeric value indicating the stop coordinate of the plot}
}
\description{
Plot the position p53 responsive elements on a DNA sequence (the mark is placed at the beginning of the responsive element)
}

