% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p53track.R
\name{p53track}
\alias{p53track}
\title{Function p53.track}
\usage{
p53track(seq, seqname = "", plot = F)
}
\arguments{
\item{seq}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}

\item{seqname}{A character string containing the name of the sequence. The default is an empty string.}

\item{plot}{A boolean value indicating whether the results should be displayed on a plot. The default is FALSE.}
}
\value{
A dataframe containing the responsive elements located on the input sequence. For each element, this infomation is provided: start and stop coordinates, spacer length, number of mismatches, sequence, mismatch string, mismatch label and grade. If plot is set to TRUE, a plot displaying the hits is also created.
}
\description{
This function locates candidate p53 responsive elements on a DNA sequence. The rules formalized in this model are manually curated and based on observations obtained during several years of experiments using the yeast-based assays, including the results presented in (Inga et al, MCB 2002; Tomso et al, PNAS 2005; Jegga et al, PNAS 2008; Jordan et al, PloS Genetics 2008; Menendez et al, PNAS 2010).
}

