\name{p53tracker-package}
\alias{p53tracker-package}
\alias{p53tracker}
\docType{package}
\title{p53tracker: detection and display of p53 binding sites on DNA sequences}
\description{p53tracker implements a pattern search algorithm that maps p53 response elements (REs) and rank them according to predicted transactivation potentials in five classes. Besides canonical, full site REs, p53tracker contains specific pattern searches for non-canonical half sites and 3/4 sites. REs can be displayed along the input sequence in a graphical layout.}
\details{
\tabular{ll}{
Package: \tab p53tracker\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-01-24\cr
License: \tab LGPL\cr
}
The main function of p53tracker is p53.track
}
\author{
Toma Tebaldi <t.tebaldi@unitn.it>
}
\references{
Literature or other references for background information
}
\keyword{ package }