
\name{SKIT-package}
\alias{SKIT}
\docType{package}
\title{
 Semiparametric Kernel Independence Test (SKIT) with excess zeros 
}
\description{
 Conduct the test of independence between two vectors when there are excess zeros. 
}
\details{
 The function \code{SKIT()} compute the test statistics and compute their \eqn{p}-values based on bootstrap samples. 
}
\references{
}
\author{Bin Zhu <bin.zhu@nih.gov> and DongHyuk Lee <donghyuk.lee@nih.gov>}
\keyword{package}