% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paili}
\alias{paili}
\title{Weekly deaths from Influenza-like illness in Pennsylvania between 1972 and 1998.}
\format{
A data frame with 1404 rows and 3 variables:
\describe{
  \item{PENNSYLVANIA}{weekly deaths}
  \item{YEAR}{the year}
  \item{WEEK}{the week}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
paili
}
\description{
A dataset containing the weekly ILI 
related deaths in  Pennsylvania between 1972 and 1998.
}
\keyword{datasets}
