% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palymes}
\alias{palymes}
\title{Weekly incidence of Lymes disease in Pennsylvania between 2006 and 2014.}
\format{
A data frame with 448 rows and 3 variables:
\describe{
  \item{PENNSYLVANIA}{weekly incidence}
  \item{YEAR}{the year}
  \item{WEEK}{the week}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
palymes
}
\description{
A dataset containing the weekly incidence of Lymes disease
 in  Pennsylvania between 2006 and 2014.
}
\keyword{datasets}
