% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pameasle}
\alias{pameasle}
\title{Weekly incidence of measles in Pennsylvania between 1928 and 1969.}
\format{
A data frame with 448 rows and 3 variables:
\describe{
  \item{PENNSYLVANIA}{weekly incidence}
  \item{YEAR}{the year}
  \item{WEEK}{the week}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
pameasle
}
\description{
A dataset containing the weekly incidence of measles
 in  Pennsylvania between 2006 and 2014.
}
\keyword{datasets}
