
\name{PatInfo}
\alias{PatInfo}

%\docType{package}
\title{
  Data for the \code{\link{MutScotPatientHeter}} example
}
\description{
 Data for the \code{\link{MutScotPatientHeter}} example.
}
\details{ 
  The object \code{data} is a data frame with columns for the sample id and ER status.
}
\examples{
  data(PatInfo, package="MUTSCOT")
  PatInfo[1:5, ]
}
\keyword{data}
%\seealso{ \code{\link{subhmm}} }



