% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdv}
\alias{pdv}
\title{Cumulative death count of harbor seals from CDV.}
\format{
A list with three items:
\describe{
\item{coord}{A data frame with 3 columns and 25 rows. Column location 
represents location name, latitude is latitude and longitude is longitude}
  \item{ts}{A data frame with 26 columns and 269 rows. The first column is the day 
 since beginning of outbreak and the next 25 columns are cumulative count of stranded
seal carcasses}
\item{fs}{A 25-by-25 matrix representing the seaway friction distance among the haulouts.}
}
}
\source{
Harding, K. C., Härkönen, T. and Caswell, H. (2002), The 2002 European seal plague: epidemiology and population consequences. Ecology Letters, 5: 727-732. <doi:10.1046/j.1461-0248.2002.00390.x>
}
\usage{
pdv
}
\description{
The cumulative count of dead seals washed ashore across
25 Northern European areas during the 2002 epidemic starting in
May and running through the end of the year.
}
\keyword{datasets}
