% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pertcop}
\alias{pertcop}
\title{Weekly whooping cough incidence from 1900-1937 in Copenhagen, Denmark.}
\format{
A data frame with 1982 rows and 9 variables:
\describe{
  \item{date}{date}
  \item{births}{births}
  \item{day}{day of month}
  \item{month}{month of year}
  \item{year}{year}
  \item{cases}{weekly incidence }
  \item{deaths}{weekly deaths}
  \item{popsize}{weekly population size interpolated from census data}
}
}
\source{
Lavine et al. 2013. Immune boosting explains regime- shifts in prevaccine-era pertussis dynamics. PLoS ONE, 8(8):e72086. <doi:10.1371/journal.pone.0072086>
}
\usage{
pertcop
}
\description{
A dataset containing the weekly incidence of whooping cough from
Copenhagen, Denmark between January 1900 and December 1937
}
\keyword{datasets}
