% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peru}
\alias{peru}
\title{Rubella in Peru data.}
\format{
A data frame with 95 rows and 2 variables:
\describe{
  \item{age}{end of age-bracket (in years)}
  \item{cumulative}{cumulative number of rubella cases}
  \item{incidence}{number of rubella cases}
  \item{n}{total cases}
}
}
\source{
Metcalf et al (2011) Rubella metapopulation dynamics and importance of spatial coupling to the risk of congenital rubella syndrome in Peru. Journal of the Royal Society Interface 8: 369-376. \url{http://doi:10.1371/journal.pone.0072086}
}
\usage{
peru
}
\description{
Rubella incidence by age as studied by Metcalf et al (2011).
}
\keyword{datasets}
