% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexDistribution.R
\name{plotSexDistribution}
\alias{plotSexDistribution}
\title{A plotSexDistribution Function}
\usage{
plotSexDistribution(beta.value = NULL, detecP = NULL,
  beta.intervals.X = seq(0, 1, 0.1), beta.intervals.Y = seq(0, 1, 0.1),
  p.intervals.X = c(-18, -5, -2, 0), p.intervals.Y = c(-18, -5, -2, 0),
  samples = NULL, filename = NULL, include_reference = TRUE,
  color = NULL)
}
\arguments{
\item{detecP}{detection p-value matrix}

\item{beta.intervals.X}{Intervals of beta-value of X chromosome probes within each of which the proportion is calculated. Default is 'seq(0,1,0.1)', which is 0 to 0.1, 0.1 to 0.2, ..., 0.9 to 1.0.}

\item{beta.intervals.Y}{Intervals of beta-value of Y chromosome probes within each of which the proportion is calculated. Default is the same as that of 'beta.intervals.X'}

\item{p.intervals.X}{Intervals of detection p-value of X chromosome probes within each of which the proportion is calculated. Default is -Inf to 1e-18m 1e-18 to 1e-5, 1e-5 to 1e-2, 1e-2 to 0.}

\item{p.intervals.Y}{Intervals of detection p-value of Y chromosome probes within each of which the proportion is calculated. Default is the same as that of 'p.intervals.X'}

\item{samples}{the list of samples to be plotted. If NULL (default) all samples in the beta.value matrix are plotted.}

\item{filename}{When specified, a PDF is generated.}

\item{include_reference}{Logical value to specify whether to plot the reference data  (Default is FALSE).}

\item{color}{Logical value to specify whether to plot the reference data  (Default is '#984ea3', one of the shades of purple).}

\item{beta-value}{beta-value matrix}

\item{main}{character string to be used as the plot title. Default is an empty string.}
}
\description{
This function generates a line chart using 'ggplot' showing the distribution of beta-values of chrX and chrY and detection p-values of chrY.
}
\examples{
plotSexDistribution(beta.value=beta, detecP=pval)
}
\keyword{plotSexDistribution}
