% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexEstimation.R
\name{plotSexEstimation}
\alias{plotSexEstimation}
\title{A plotSexEstimation Function}
\usage{
plotSexEstimation(sex_estimation = NULL, samples = NULL, main = "",
  filename = NULL, include_reference = FALSE)
}
\arguments{
\item{sex_estimation}{result of estimateSex function}

\item{samples}{the list of samples to be plotted. If NULL (default) all samples in the estimateSex result are plotted.}

\item{main}{character string to be used as the plot title. Default is an empty string.}

\item{filename}{When specified, a PDF is generated.}

\item{include_reference}{Logical value to specify whether to plot the reference data  (Default is FALSE).}
}
\description{
This function generates a scatter plot using 'ggplot' showing the first principal component of two principal component analysis results (pincipal component analysis using the data from X chromosome and the data from Y chromosome)
}
\examples{
plotSexEstimation(sex_estimation = sexEst)
}
\keyword{plotSexEstimation}
