% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{plotUnivariate}
\alias{plotUnivariate}
\title{Plot univariate feature selection object.}
\usage{
plotUnivariate(results, times, conds, rep, path = ".",
  pdf.name = "topResults.pdf", top = 100)
}
\arguments{
\item{results}{the output list from the featureSelectionHMM/featureSelectionOneConditionHMM function}

\item{times}{vector with all the time-points}

\item{conds}{vector with the condition names}

\item{rep}{Number of biological replicates}

\item{path}{Directory path to set the pdf. Defaults to '.'}

\item{pdf.name}{Character - pdf name. Defaults to 'topResults.pdf'}

\item{top}{Number of genes to be plotted. Defaults to 100}
}
\description{
This function plots in a pdf file the top ranked features stored in a
feature selection object that summarized the biological replicates 
via a statistical measure.
}
\examples{

data("ratTimeSeries")

# All the necessary variables to run the script
times <- c(0,2,6,12,18,24)  # study time-points vector
conds <- c("CTRL","IKAROS") # Condition vector
reps <- 3                   # number of biological replicates
\dontrun{
# It has some random sampling inside, so a set.seed is recommended
# Case: Biological replicates summarized by median
set.seed(100)
summ.results <- featureSelectionHMM(data.ann,     # Gene expression data matrix
                                   times,         # time-point vector
                                   conds,         # Condition vector
                                   reps,          # Number of replicates
                                   USE.REP  = F,  # Do not use all replicate data
                                   func = median  # Summarize the data 
)  

plotUnivariate(summ.results, times, conds[2], reps, ".", "Fat.Rat.uni.res.pdf")
}

}
\seealso{
\code{\link{plotMultivariate}} for not summarized biological replicates.
}
\keyword{FS,}
