% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polymod}
\alias{polymod}
\title{POLYMOD contact-rate data by Age.}
\format{
A data frame with 900 rows and 3 variables:
\describe{
  \item{contactor}{end of age-bracket (in years) of contactor group}
  \item{contactee}{end of age-bracket (in years) of contactee group}
  \item{contact.rate}{average contact rate}
}
}
\source{
Mossong et al. 2008 Social contacts and mixing patterns relevant to the spread of infectious diseases PLoS Med, Public Library of Science  5:e74. <doi:10.1371/journal.pmed.0050074>.
}
\usage{
polymod
}
\description{
Age-specific contact rates from the diary study by Mossong et al. 2008.
}
\keyword{datasets}
