% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{puc.rank.plot}
\alias{puc.rank.plot}
\title{puc.rank.plot is a ggplot scatter plot with desired srcMeans variable on the y axis and chemical rank on the x axis.
The dependent variable on the y axis can be changed using output.variable argument, but only one output variable can be plotted at a time.
User can either plot the individual product or the higher level puc using the combine argument. Data comes from all_srcMeans output file, so data points are the mean for that output variable.}
\usage{
puc.rank.plot(run.name, output.variable = "exp.dermal", label_chem = FALSE)
}
\arguments{
\item{run.name}{default = name of last run (in current R session)}

\item{output.variable}{argument is a single character string. default = "exp.dermal".}
}
\value{
a plot of output.variable vs chemical rank with a legend for color and shape
}
\description{
puc.rank.plot is a ggplot scatter plot with desired srcMeans variable on the y axis and chemical rank on the x axis.
The dependent variable on the y axis can be changed using output.variable argument, but only one output variable can be plotted at a time.
User can either plot the individual product or the higher level puc using the combine argument. Data comes from all_srcMeans output file, so data points are the mean for that output variable.
}
\details{
A SHEDS run creates an output file for each chemical. This plot pulls from the srcMeans.csv for each chemical in the output file.
Chemicals are ranked in ascending order of output.variable on the x axis. The y axis is log transformed and a constant of 1e-10 added to y variable for
log transformation of 0's. The plot colors points by higher level PUC.
}
