% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{r0fun}
\alias{r0fun}
\title{Function to calculate R0 from a contact matrix}
\usage{
r0fun(CM, tau, gamma)
}
\arguments{
\item{CM}{an object of class CM}

\item{tau}{= probability of infection across an edge}

\item{gamma}{= probability of removal per time step}
}
\value{
the R0
}
\description{
Function to calculate R0 from a contact matrix
}
\examples{
cm1=barabasiAlbert(N=200,K=2)
r0fun(cm1, 0.3, 0.1)
}
