% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit}
\alias{rabbit}
\title{Rabbit \emph{Bordetella bronchiseptica} data.}
\format{
A data frame with 42 rows and 3 variables:
\describe{
  \item{a}{end of age-bracket (in months)}
  \item{n}{number of rabbits tested}
  \item{inf}{number of rabbits infected with the bacterium}
}
}
\source{
Long et al (2010) Identifying the Age Cohort Responsible for Transmission in a Natural Outbreak of Bordetella bronchiseptica. PLoS Pathogens 6(12): e1001224. \url{http://doi:10.1371/journal.ppat.1001224}
}
\usage{
rabbit
}
\description{
Rabbits infected by \emph{B. bronchiseptica} by age as studied by Long et al (2010).
}
\keyword{datasets}
