% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabies}
\alias{rabies}
\title{Raccoon rabies data.}
\format{
A data frame with 208 rows and 12 variables:
\describe{
  \item{Month}{Month since rabies appearance in the state}
  \item{CT}{Connecticut}
  \item{DE}{Delaware}
  \item{MD}{Maryland}
  \item{MA}{Massachusetts}
  \item{NJ}{New Jersey}
  \item{NY}{New York}
  \item{NC}{North Carolina}
  \item{PA}{Pennsylvania}
  \item{RI}{Rhode Island}
  \item{VA}{Virginia}
  \item{WV}{West Virginia}
}
}
\source{
Childs et al. 2000. Predicting the local dynamics of epizootic rabies among raccoons in the United States Proceedings of the National Academy of Sciences 97:13666-13671. <doi:10.1073/pnas.240326697>
}
\usage{
rabies
}
\description{
Data is the average monthly number of reported cases of rabid raccoons across all counties within each of 
11 east coast US states the time line is from the first reported case in each state (starting in late 1970s for West Virginia).
}
\keyword{datasets}
