% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{read.chem.props}
\alias{read.chem.props}
\title{read.chem.props}
\usage{
read.chem.props(filename, specs)
}
\description{
Read.chem.props reads the chemical properties from the .csv file indicated by "filename". "Specs" contains the run specifications from the run.file, and is used to subset the chemicals by the list provided on the run.file.  If no such list was given, then all chemicals are kept.
}
