% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{read.exp.factors}
\alias{read.exp.factors}
\title{read.exp.factors}
\usage{
read.exp.factors(filename)
}
\description{
Read.exp.factors reads the exposure factors from the .csv file indicated by "filename".
}
