% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{read.run.file}
\alias{read.run.file}
\title{read.run.file}
\usage{
read.run.file(run.file = "run_test.txt")
}
\description{
Each SHEDS run has its own "run.file" that the user prepares before the run.   This file contains all the settings and file references needed for the run.  Read.run.file occurs at the start of each SHEDS run.  The contents of the run.file are examined, and stored in the R object "specs".
}
