% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{read.source.scen.file}
\alias{read.source.scen.file}
\title{read.source.scen.file}
\usage{
read.source.scen.file(filename)
}
\description{
Read.source.scen.file reads the list of active exposure scenarios for each potential source of chemical from the .csv file indicated by "filename".
}
