% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tdtk.R, R/read_trauma.R
\name{read_tdtk}
\alias{read_tdtk}
\title{Read Trauma is a Fucnction to read in trauma data}
\usage{
read_tdtk(file_name = ..., file_type = "csv", save_file = FALSE,
  blind = NA, extract_ts = FALSE, random = TRUE, warning = FALSE)

read_tdtk(file_name = ..., file_type = "csv", save_file = FALSE,
  blind = NA, extract_ts = FALSE, random = TRUE, warning = FALSE)
}
\arguments{
\item{file_name}{is the name of the file}

\item{file_type}{is the type of file ("csv" or "xls" for now with
"SQL" functionality to be added later.}

\item{save_file}{is T/F and will save the file prior to returnign
the dataframe/list}

\item{blind}{is a case type of blinding for ("NA", "zip", "age" or
"both")}

\item{extract_ts}{is a T/F to extract the time series data (if
included) and will return a list wtih TS and PT data}

\item{random}{is the shuffling of unblinded data prior to use
blindied data is automatically shuffled once time series data
is removed.}

\item{warning}{T/F to turn off warnings for unblinded data}

\item{save_data}{meant to be used as a T/F for saving data}

\item{file_name}{is the name of the file}

\item{file_type}{is the type of file ("csv" or "xls" for now with
"SQL" functionality to be added later.}

\item{save_file}{is T/F and will save the file prior to returnign
the dataframe/list}

\item{blind}{is a case type of blinding for ("NA", "sip", "age" or
"both")}

\item{random}{is the shuffling of unblinded data prior to use
blindied data is automatically shuffled once time series data
is removed.}

\item{warning}{T/F to turn off warnings for unblinded data}
}
\value{
either a dataframe extract_ts = T will return a timeseries

either a dataframe or a list if extract_ts = T
}
\description{
Given either CSV or exl file with standard column names and order
be able to read in the file, extract time-series data (if
included), clean zip codes, determine ISS catagory and blind that
data (IFF blidn_data == TRUE).

Given either CSV or exl file with standard column names and order
be able to read in the file, extract time-series data (if
included), clean zip codes, determine ISS catagory and blind that
data (IFF blidn_data == TRUE).
}
\details{
The read_trauma() function was designed to work with exported data
from an electronic health record source.  As the function continues
to develop it is expected to add direct SQL backend queeries to
minimize the need to export.  This is probably an inefficient way
to load the file etc but its a start.

The read_trauma() function was designed to work with exported data
from an electronic health record source.  As the function continues
to develop it is expected to add direct SQL backend queeries to
minimize the need to export.
}
\examples{


}
\references{
\url{https:\\www.github.com\eric43\tdtk}
}
\seealso{
\code{\link{<tdtk_read.r>}}
}
\author{
Eric W. Olle, \email{eric.olle@gmail.com}

Eric W. Olle, \email{eric.olle@gmail.com}
}
