% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_retixmp.R
\name{retixmp}
\alias{retixmp}
\title{ixmp core}
\description{
This page provides documentation for the ixmp-R implementation
}
\details{
Main features:

- a platform core linked to a database instance,

- integration with the |MESSAGEix| Integrated Assessment model
   and other numerical models
}
