% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\docType{data}
\name{ricker.app}
\alias{ricker.app}
\title{Launch a shiny-app simulating the Ricker model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
ricker.app
}
\description{
Launch a shiny-app simulating the Ricker model
}
\details{
Launch app for details
}
\examples{
\dontrun{Ricker.app}
}
\keyword{datasets}
