% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{ringlattice}
\alias{ringlattice}
\title{Function to generate a ring lattice}
\usage{
ringlattice(N, K)
}
\arguments{
\item{N}{the number of nodes}

\item{K}{the number of neighbors to which each node is connected so degree = 2xK}
}
\value{
An object of class CM (contact matrix)
}
\description{
Function to generate a ring lattice
}
\examples{
cm=ringlattice(N=20,K=4)
}
