% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyross.r
\docType{data}
\name{ross.app}
\alias{ross.app}
\title{Launch a shiny-app simulating a Ross-Macdonald model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
ross.app
}
\description{
Launch a shiny-app simulating a Ross-Macdonald model
}
\details{
Launch app for details.
}
\examples{
\dontrun{ross.app}
}
\keyword{datasets}
