% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{rossmacdonald.app}
\alias{rossmacdonald.app}
\title{Launch a shiny-app simulating the Ross-Macdonald malaria model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
rossmacdonald.app
}
\description{
Launch a shiny-app simulating the Ross-Macdonald malaria model
}
\details{
Launch app for details
}
\examples{
\dontrun{rossmacdonald.app}
}
\keyword{datasets}
