\name{rsse_batch_ui}
\alias{rsse_batch_ui}
\title{rsse_batch_ui}
\usage{
rsse_batch_ui()
}
\description{
A function to display user friendly interface to do
estitamete the sample size for differential expression
analysis of RNA-seq data at batch mode. The result can be
saved to file in csv format.
}
\examples{
\dontrun{
rsse_batch_ui()
}
}

