% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShedsHT.R
\name{run}
\alias{run}
\title{run}
\usage{
run(run.file = "", wd = "")
}
\arguments{
\item{run.file}{The name of the run file to be used for a given run.  Many different "Run"" files may be set up for special
purposes.The one being invoked must be present in the inputs folder.}

\item{wd}{The user's working directory. The working directory should contain an inputs folder, containing all necessary 
SHEDS.HT input files. The wd value should be set once for each SHEDS.HT installation by replacing the default value in
the definition of run().}
}
\value{
No variable will be returned.
}
\description{
Function to call the \code{\link{Run}} txt file, which consists of user-defined parameters and calls to input files required 
to initialize a SHEDS.HT run.
}
\details{
This function is used to provide R with information necessary to produce a SHEDS-HT run and to initialize necessary
parameters. No values are returned. In order to produce a successful run, all necessary input files should be stored in the
 working directory specified by the wd argument.
}
\seealso{
\code{\link{Run.txt}}, \code{\link{read.run.file}}
}
\author{
Kristin Isaacs, Graham Glen
}
