% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gams.R
\name{run_gams}
\alias{run_gams}
\title{Auxiliary functions for executing GAMS}
\usage{
run_gams(model_file, args, gams_args = "LogOption=4")
}
\arguments{
\item{model_file}{the path to the gams file}

\item{args}{arguments related to the GAMS code (input/output gdx paths, etc.)}

\item{gams_args}{additional arguments for the CLI call to gams
`LogOption=4` prints output to stdout (not console) and the log file}
}
\description{
Auxiliary functions for executing GAMS
}
