
\name{SALMON-package}
\alias{SALMON}
\docType{package}
\title{
    Signature Activity and Expectancy
}
\description{
  To estimate signature activities and calculate signature expectancy.
}
\details{
The two main functions in this package are \code{\link{EstimateSigActivity}} and 
\code{\link{CalculateSigExpectancy}}.
}
\references{
Lee, D., Wang, D., Yang, X., Shi, J., Landi, M., Zhu, B. (2021) 
SUITOR: selecting the number of mutational signatures through cross-validation.
bioRxiv, doi: https://doi.org/10.1101/2021.07.28.454269. 
}
\author{Donghyuk Lee <dhyuklee@pusan.ac.kr> and Bin Zhu <bin.zhu@nih.gov>}
\keyword{package}