% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyseir.R
\docType{data}
\name{seir.app}
\alias{seir.app}
\title{Launch a shiny-app simulating the seasonal SEIR model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
seir.app
}
\description{
Launch a shiny-app simulating the seasonal SEIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{seir.app}
}
\keyword{datasets}
