% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyseirs.R
\docType{data}
\name{seirs.app}
\alias{seirs.app}
\title{Launch a shiny-app simulating the SEIRS model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
seirs.app
}
\description{
Launch a shiny-app simulating the SEIRS model
}
\details{
Launch app for details
}
\examples{
\dontrun{seirs.app}
}
\keyword{datasets}
