% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{set_hash}
\alias{set_hash}
\title{Set a Hash Attribute on an Object}
\usage{
set_hash(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The modified object.
}
\description{
Set a Hash Attribute on an Object
}
\seealso{
Other hash functions for objects: 
\code{\link{un_hash}()}
}
\concept{hash functions for objects}
