\name{ setnv }
\alias{ setnv }
\title{set normative values and location map}
\description{
set normative values and stimuli location map to be used with visualFields
}
\usage{setnv( nvtxt = "nvsapdefault" )}
\arguments{
  \item{nvtxt}{char. Name of the normative reference values to be used. It must be an existing structure}
}
\details{
sets normative values and stimuli location map to be used with visualFields
}

\seealso{ \code{\link{getnv}} }

\examples{
  setnv()
}

\author{Ivan Marin-Franch}