% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfreports.R
\name{vfsfa}
\alias{vfsfa}
\title{Single Field Analysis}
\usage{
vfsfa(vf, file, ...)
}
\arguments{
\item{vf}{visual field data}

\item{file}{The pdf file name where to save the single visual field
analysis report}

\item{...}{other graphical arguments}
}
\description{
Performs a single field analysis, including TD and PD
values and probability maps and global indices and their p-values.
}
\details{
\itemize{
  \item\code{vfsfa} saves a pdf with a visual field statistical report
  \item\code{vfsfashiny} reports the results interactively with shiny
}
}
