% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sh9}
\alias{sh9}
\title{Daily measures of malaria infected mice.}
\format{
A data frame with 1300 rows and 11 variables:
\describe{
  \item{Line}{line number}
  \item{Day}{day of infection}
  \item{Box}{Cage number}
  \item{Mouse}{Mouse identifier}
  \item{Treatment}{Plasmodium strain}
  \item{Ind2}{Unique mouse identifier}
  \item{Weight}{Mouse weight}
  \item{Glucose}{Blood glucose level}
  \item{RBC}{Red blood cell count}
  \item{Sample}{Sample number}
  \item{Para}{Parasite count}
}
}
\source{
Sylvie Huijben
}
\usage{
sh9
}
\description{
Daily data on laboratory mice infected with various strains of \emph{Plasmodium chabaudi}
}
\keyword{datasets}
