% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{sim.cb}
\alias{sim.cb}
\title{Function to simulate the chain-binomial model}
\usage{
sim.cb(S0, beta)
}
\arguments{
\item{S0}{a scalar with value for S0}

\item{beta}{a scalar with value for beta}
}
\value{
A data-frame with time series of susceptibles and infecteds
}
\description{
Function to simulate the chain-binomial model
}
\examples{
sim=sim.cb(S0=6500, beta=23)
}
