% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysir.R
\docType{data}
\name{sir.app}
\alias{sir.app}
\title{Launch a shiny-app simulating the SIR model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
sir.app
}
\description{
Launch a shiny-app simulating the SIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{sir.app}
}
\keyword{datasets}
