% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{sivmod}
\alias{sivmod}
\title{Gradient-function for the SIR model with outbreak-response vaccination}
\usage{
sivmod(t, x, parms)
}
\arguments{
\item{t}{Implicit argument for time}

\item{x}{A vector with values for the states}

\item{parms}{A vector with parameter values for the SIR system}
}
\value{
A list of gradients
}
\description{
Gradient-function for the SIR model with outbreak-response vaccination
}
\seealso{
\code{\link{retrospec}}
}
