% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfreports.R
\name{vfspa}
\alias{vfspa}
\title{Series Progession Analysis}
\usage{
vfspa(
  vf,
  file,
  type = "td",
  nperm = factorial(7),
  trunc = 1,
  testSlope = 0,
  ...
)
}
\arguments{
\item{vf}{A \code{VisualField} object}

\item{file}{The pdf file name where to save the single visual field
analysis report. If `\code{FALSE}` (default) it shows an interactive
report with Shiny}

\item{type}{Type of data to use. It can be `\code{s}`, `\code{td}`, or
`\code{pd}`.}

\item{nperm}{Number of permutations. Default is 7!}

\item{trunc}{value for the Truncated Product Method (see reference).
Default is 1}

\item{testSlope}{slope, or slopes, to test as null hypothesis. Default is 0.
if a single value, then the same null hypothesis is used for all locations.
If a vector of values, then (for \code{plr} and \code{poplr}) each
location of the visual field will have a different null hypothesis. The length
of testSlope must be 1 or equal to the number of locations to be used in the PLR
or PoPLR analysis}

\item{...}{other graphical arguments}
}
\description{
Performs a series progression analysis, including PoPLR
(see reference) and the global indices Mean Sensitivity, Mean Deviation,
and General Height and reports the results graphically either
interactively in a browser (with Shiny) or in a pdf file.
}
\references{
N. O'Leary, B. C. Chauhan, and P. H. Artes. \emph{Visual field progression in
glaucoma: estimating the overall significance of deterioration with permutation
analyses of pointwise linear regression (PoPLR)}. Investigative Ophthalmology
and Visual Science, 53, 2012
}
