% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{subset_sizes}
\alias{subset_sizes}
\title{Determine Subset Sizes Close to Equality}
\usage{
subset_sizes(n, k)
}
\arguments{
\item{n}{The size of the set.}

\item{k}{The number of subsets.}
}
\value{
A vector of k sizes of the subsets.
}
\description{
Determine the sizes of k subsets of a set with
n elements in such a way that the sizes are as
equal as possible.
}
\examples{
subset_sizes(n = 100, k = 6)
subset_sizes(n = 2, k = 6)
}
\seealso{
Other subsetting functions: 
\code{\link{index_groups}()}
}
\concept{subsetting functions}
