
\name{SUITOR-package}
\alias{SUITOR}
\docType{package}
\title{
 Selecting the number of mutational signatures through cross-validation
}
\description{
 Selecting the number of mutational signatures through cross-validation
}
\details{
 
}
\references{
 
}
\author{Donghyuk Lee <donghyuk.lee@nih.gov> and Bin Zhu <bin.zhu@nih.gov>}
\keyword{package}