% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{summarize.chemical}
\alias{summarize.chemical}
\title{summarize.chemcial}
\usage{
summarize.chemical(x, c, chem, chemical, set, sets, specs)
}
\arguments{
\item{x}{Default = none        Exposure data set}

\item{c}{Default = none    Index # for chemical}

\item{chem}{Default = none    CAS for chemical}

\item{chemical}{Default = none    Full chemical name}

\item{set}{Default = none    Index # for set of simulated persons}

\item{sets}{Default = none        Total # of sets in this SHEDS run}

\item{specs}{Default = none List of settings from the "run" input file}
}
\value{
If "x" is a single set of data (that is, if the argument "set" is between 1 and sets, inclusive), then the .csv
file created by this function has the suffix "_set#stats.csv", where "#" is the set number.  If the "set" argument is
"allstats", then the .csv file has the suffix _allstats.csv".  The output file contains the tables for all cohorts with a
non-zero population.
}
\description{
Summarize.chemical writes a .csv file containing a summary of the exposure and dose results from the object "x".
}
\details{
Tables are produced for each of the following cohorts: males, females, females ages 16-49, age 0-5, age 6-11,
age 12-19, age 20-65, age 66+, and a table for all persons. The variables that are summarized are: dermal exposure,
ingestion exposure, inhalation exposure, inhaled dose, intake dose, dermal absorption, ingestion absorption, inhalation
absorption, total absorption in micrograms per day, total absorption in milligrams per kilogram per day, and chemical mass
down the drain.
In each table, the following statistics are computed across the appropriate subpopulation: mean, standard deviation, and
quantiles .005, .01, .025, .05, .1, .15, .2, .25, .3, .4, .5, .6, .7, .75, .8, .85, .9, .95, .975, .99, .995.
}
\author{
Kristin Isaacs, Graham Glen
}
