% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{summary.cm}
\alias{summary.cm}
\title{Function to calculate the degree distribution for an object of class CM}
\usage{
\method{summary}{cm}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{an object of class cm}

\item{plot}{if TRUE a bar plot of the degree distribution is produced}

\item{...}{other arguments}
}
\value{
A plot of the contract matrix
}
\description{
Function to calculate the degree distribution for an object of class CM
}
\examples{
cm=wattsStrogatz(N=20, K=4, Prw=.3)
summary(cm)
}
