% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{summary.netSir}
\alias{summary.netSir}
\title{Function to summarize a netSir object}
\usage{
\method{summary}{netSir}(object, ...)
}
\arguments{
\item{object}{an object of class netSir}

\item{...}{other arguments}
}
\value{
A data-frame with the time series of susceptible, infected and recovered individuals
}
\description{
Function to summarize a netSir object
}
\seealso{
\code{\link{sirNetmod}}
}
