% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw.R
\name{throw}
\alias{throw}
\title{Throw a Condition}
\usage{
throw(message_string, system_call = sys.call(-1), ...)
}
\arguments{
\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}

\item{...}{Arguments to be passed to
\code{\link[base:structure]{base::structure}}.}
}
\value{
The function does never return anything, it stops with a
condition of class c("error", "fritools", "condition").
}
\description{
Throws a condition of class c("error", "fritools", "condition").
}
\details{
We use this condition as an error dedicated to \pkg{ fritools.}
}
\seealso{
Other bits and pieces: 
\code{\link{golden_ratio}()},
\code{\link{is_difftime_less}()},
\code{\link{is_valid_primary_key}()},
\code{\link{r_cmd_install}()},
\code{\link{str2num}()},
\code{\link{strip_off_attributes}()},
\code{\link{tapply}()}
}
\concept{bits and pieces}
\keyword{internal}
