% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{trimzero}
\alias{trimzero}
\title{Trimzero}
\usage{
trimzero(x, y)
}
\arguments{
\item{x}{aCAS number. Defult is none}

\item{y}{A dummy argument This helps with the removal of the zeros inteh CAS number}
}
\value{
y a shorter CAS number. If the initial part is all zero (as in "0000000_12_3"), one zero is left in the first part
(that is, "0_12_3" for this example).
}
\description{
This function removes initial zeroes from CAS numbers.
}
\details{
Each CAS number has three parts, separated by underscores.  The first part is up to seven digits, but optionally,
leading zeroes are omitted.  For example, formaldehyde may be either "0000050_00_0" or "50_00_0". SHEDS needs to match
CAS numbers across input files, and trimzero is used to ensure matching even when the input files follow different conventions.
}
\author{
Kristin Isaacs, Graham Glen
}
