% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny2.r
\docType{data}
\name{tsir.app}
\alias{tsir.app}
\title{Launch a shiny-app simulating TSIR model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
tsir.app
}
\description{
Launch a shiny-app simulating TSIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{tsir.app}
}
\keyword{datasets}
