% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsir.R
\name{tsirLlyap}
\alias{tsirLlyap}
\title{Function to calculate the local Lyapunov exponents for the TSIR}
\usage{
tsirLlyap(x, m = 1)
}
\arguments{
\item{x}{an object of class \code{lyap} (normally from a call to \code{tsirLyap})}

\item{m}{number of forward iterations on the attractor}
}
\value{
An object of class llyap with the local Lyapunov exponent and S-I data
}
\description{
Function to calculate the local Lyapunov exponents from an object of class \code{lyap}.
}
\examples{
\dontrun{see chapter 10 in book}
}
