% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny2strain.r
\docType{data}
\name{twostrain.app}
\alias{twostrain.app}
\title{Launch a shiny-app simulating a two-strain SIR model}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
twostrain.app
}
\description{
Launch a shiny-app simulating a two-strain SIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{twostrain.app}
}
\keyword{datasets}
