% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostrain.R
\name{twostrain}
\alias{twostrain}
\title{Gradient-function for the twostrain model}
\usage{
twostrain(t, y, parameters)
}
\arguments{
\item{t}{Implicit argument for time}

\item{y}{A vector with initial values for the states}

\item{parameters}{A vector with parameter values for the system}
}
\value{
A list of gradients
}
\description{
Gradient-function for the twostrain model
}
\examples{
require(deSolve)
times  = seq(0, 5, by=1/100)
paras  = c(mu = 5, N = 1, R01=2, R02=1.5, gamma = 365/14, Theta=0.9, Xi=0.9, Pi=0.9)
start = c(S=0.998, I1=0.001, I2=0.001,R1 = 0, R2=0, J1=0, J2=0, R=0)
out=as.data.frame(ode(start, times, twostrain, paras))
}
