% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tyscarlet}
\alias{tyscarlet}
\title{Weekly incidence of scarlet fever in Philadelphia between 1914 and 1947.}
\format{
A data frame with 1774 rows and 4 variables:
\describe{
  \item{YEAR}{the year}
  \item{WEEK}{the week}
  \item{PHILADELPHIA}{weekly scarlet fever incidence}
  \item{TIME}{the time counter}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
tyscarlet
}
\description{
A dataset containing the weekly incidence 
incidence of scarlet fever in Philadelphia between 1914 and 1947.
}
\keyword{datasets}
