% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tywhooping}
\alias{tywhooping}
\title{Weekly incidence of whooping cough in Philadelphia between 1925 and 1947.}
\format{
A data frame with 1200 rows and 5 variables:
\describe{
  \item{YEAR}{the year}
  \item{WEEK}{the week}
  \item{PHILADELPHIA}{weekly whooping cough incidence}
  \item{TIME}{the time counter}
  \item{TM}{observation counter}
}
}
\source{
\url{https://www.tycho.pitt.edu}
}
\usage{
tywhooping
}
\description{
A dataset containing the weekly incidence 
incidence of whooping cough in Philadelphia between 1925 and 1947.
}
\keyword{datasets}
