% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{un_hash}
\alias{un_hash}
\title{Separate an Object from its Hash Attribute}
\usage{
un_hash(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A list containing the object and its hash attribute.
}
\description{
We calculate a hash value of an object and store it as an attribute of the
objects, the hash value of that object will change. So we need to split the
hash value from the object to see whether or not the objected changed.
}
\seealso{
Other hash functions for objects: 
\code{\link{set_hash}()}
}
\concept{hash functions for objects}
