% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us}
\alias{us}
\title{2005 US Life table.}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{a}{end of age-bracket (in years)}
  \item{la}{fraction of birth cohort still alive}
  \item{fa}{fecundity at age}
  \item{sa}{survival probabilities per age-bracket}
}
}
\usage{
us
}
\description{
Survivorship and fecundities for the US in 2005 by 5 year age-brackets.
}
\keyword{datasets}
