% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usflu}
\alias{usflu}
\title{US 1975/76 ILI data.}
\format{
A data frame with 49 rows and 7 variables:
\describe{
  \item{State}{State number}
  \item{Acronym}{State code}
  \item{Pop}{Population size}
  \item{Latitude}{Latitude}
  \item{Longitude}{Longitude}
  \item{Start}{Week of start of epidemic}
  \item{Peak}{Week of peak of epidemic}
}
}
\source{
Viboud C, Bjornstad ON, Smith DL, Simonsen L, Miller MA, Grenfell BT (2006) Synchrony, waves, and spatial hierarchies in the spread of influenza. Science 312: 447-451.<doi:10.1126/science.1125237>
}
\usage{
usflu
}
\description{
Influenza-like illness data for the lower 48 states and the District of Columbia during the 1975/76 
season dominated by A/H3N2/Victoria strain
}
\keyword{datasets}
