% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variants}
\alias{variants}
\title{US SARS-CoV-2 variant data.}
\format{
A data frame with 47 rows and 7 variables:
\describe{
\item{date}{End of week of sample}
  \item{other}{Early variants}
  \item{B.1.617.2}{Delta variant}
  \item{B.1.1.529}{First omicron variant}
  \item{BA.1.1}{Omicron variant BA.1}
  \item{BA.2}{Omicron variant BA.2 and BA.2.12}
  \item{BA.2.12.1}{Omicron variant BA 2.12.1}
}
}
\source{
\url{https://coronavirus.health.ny.gov/covid-19-variant-data}
}
\usage{
variants
}
\description{
Weekly fraction of identification of the various CoV-2 variants May 2021 through March 2022.
}
\keyword{datasets}
