\name{vf91016csp1vf}
\alias{vf91016csp1vf}
\docType{data}
\title{a vf-object with CSP sample data with}
\description{a \code{\link{vf}}-object with sample data for the examples in \code{visualFields}'s help. This is real data for the right eye, but the ages have been changed to protect anonymity of the subject}
\usage{data( vf91016csp1vf )}
\format{the format is as in \code{\link{vf}} but for CSP format (see \code{\link{csplocmap}})}
\seealso{\code{\link{vf}}}

\keyword{datasets}

\author{Ivan Marin-Franch, William H Swanson, Victor E Malinovsky}