\name{vf91016right}
\alias{vf91016right}
\docType{data}
\title{a vf-object with SAP sample data}
\description{a \code{\link{vf}}-object with sample data for the examples in \code{visualFields}'s help. This is real data for the right eye, but the ages have been changed to protect anonymity of the subject}
\usage{data( vf91016right )}
\format{the format is as explained in \code{\link{vf}} with columns L1 .. L54 containing sensitivity thresholds}

\seealso{\code{\link{vf}}}

\keyword{datasets}

\author{Ivan Marin-Franch, William H Swanson, Victor E Malinovsky}