\name{vfArtes2014}
\alias{vfArtes2014}
\docType{data}
\title{Short-term retest static automated perimetry data}
\description{
Thirty patients recruited from the glaucoma clinics at the Queen Elizabeth Health Sciences Centre in Halifax, Nova Scotia. Each patient underwent 12 visual fields in 12 consecutive weekly sessions.}
\usage{data( vfArtes2014 )}

\format{It is a \code{\link{vf}}-object}

\references{
[1] P. H. Artes, N. O'Leary, M. T. Nicolela, B. C. Chauhan, and D. P. Crabb. Visual field progression in glaucoma: What is the specificity of the guided progression analysis? American Academy of Ophthalmology, 121(10):2023-2027, 2014.
}

\keyword{datasets}

\author{Paul H Artes, David P Crabb}