\name{vfaverage}
\alias{vfaverage}
\title{average of vf-objects}
\description{computes the location averages of vf-object}
\usage{vfaverage( vf )}
\arguments{
  \item{vf}{a vf-object with more than 1 entry}
}
\value{
returns the location average of vf.
}

\examples{
  vfaverage( vf91016left )
}

\author{Ivan Marin-Franch}