% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfctrSunyiu10d2.R
\docType{data}
\name{vfctrSunyiu10d2}
\alias{vfctrSunyiu10d2}
\title{SUNY-IU dataset of healthy eyes for 10-2 static automated perimetry}
\format{
See section \code{Structure of visual fields data} in \code{\link{vfdesc}}
}
\usage{
vfctrSunyiu10d2
}
\description{
SUNY-IU dataset of healthy eyes for 10-2 static automated perimetry.
Courtesy of William H Swanson.
}
\references{
H. J. Wyatt, M. W. Dul, and W. H. Swanson. \emph{Variability of visual
field measurements is correlated with the gradient of visual
sensitivity}. Vision Research, 47, 2007.

A. Shafi, W. H. Swanson, and M. W. Dul. \emph{Structure and Function
in Patients with Glaucomatous Defects Near Fixation}. Optometry and
Vision Science, 88, 2011.
}
\seealso{
\code{\link{vfpwgSunyiu24d2}}, \code{\link{vfctrIowaPC26}},
         \code{\link{vfctrIowaPeri}}, \code{\link{vfctrSunyiu24d2}},
         \code{\link{vfpwgRetest24d2}}
}
\keyword{dataset}
