% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfctrSunyiu24d2.R
\docType{data}
\name{vfctrSunyiu24d2}
\alias{vfctrSunyiu24d2}
\title{SUNY-IU dataset of healthy eyes for 24-2 static automated perimetry}
\format{
See section \code{Structure of visual fields data} in \code{\link{vfdesc}}
}
\usage{
vfctrSunyiu24d2
}
\description{
This dataset of healthy eyes was used to generate the normative values
\code{sunyiu_24d2}, \code{sunyiu_24d2_pw}, \code{sunyiu_24d2}, and
\code{sunyiu_24d2_pw_cps} included in \code{\link{normvals}}.
Courtesy of William H Swanson and Mitch W Dul
}
\references{
H. J. Wyatt, M. W. Dul, and W. H. Swanson. \emph{Variability of visual
field measurements is correlated with the gradient of visual
sensitivity}. Vision Research, 47, 2007.

A. Shafi, W. H. Swanson, and M. W. Dul. \emph{Structure and Function
in Patients with Glaucomatous Defects Near Fixation}. Optometry and
Vision Science, 88, 2011.
}
\seealso{
\code{\link{vfpwgSunyiu24d2}}, \code{\link{vfctrIowaPC26}},
         \code{\link{vfctrIowaPeri}}, \code{\link{vfctrSunyiu10d2}},
         \code{\link{vfpwgRetest24d2}}
}
\keyword{dataset}
