% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfevent.R
\name{vfcriteria}
\alias{vfcriteria}
\alias{hap2}
\alias{ukgts}
\alias{ght}
\alias{foster}
\alias{logts}
\title{Event criteria for detection of glaucomatous damage}
\usage{
vfcriteria(vf, criteria = "all", vf.ctrl = visualFields::vfctrSunyiu24d2)

hap2(vf, pd, pdp, ght, helper)

ukgts(vf, td, tdp, helper)

ght(vf, td, pd, pdp, lims, helper)

foster(pd, pdp, ght, helper)

logts(td, helper)
}
\arguments{
\item{vf}{Visual fields to be analyzed as a standard vfobject}

\item{criteria}{select to use all criteria (default), or a single
criteria: hap2, ukgts, ght, foster, or logts}

\item{vf.ctrl}{control data (HERMAN: DESCRIBE IT. I WOULD CHANGE THE NAME
OF THIS PARAMETER TO vfctr)}

\item{pd}{pattern deviation map to be analyzed}

\item{pdp}{pattern deviation probability map to be analyzed}

\item{ght}{result of the GHT}

\item{helper}{helper map tht specifies the surrounding points of each point}

\item{td}{total deviation map to be analyzed}

\item{tdp}{total deviation probability map to be analyzed}

\item{lims}{limits of normality required for the GHT algorithm}
}
\value{
whether analysis resulted in a detection of a defect. For
\code{vfcriteria} with the
 different criteria selected with \code{criteria}
--------------------------------------------------------------------------------------------------------------------------------------   
Calculate limits of normality as described in Asman & Heijl, Arch Opthalmol, 1992
1. pass about 100+ vfs of normal eyes, preferably with 1 unique eye per patient, within the vf_ctrl parameter
2. for each vf in vf_ctrl, calculate gh, sector sums, and sector up-down differences
3. use the bootstrap method to calculate limits of normality for ght, sector sums, and sector up-down differences
   a. select 500 random samples from vf_ctrl, making sure each sample has 1 unique eye per patient
   b. for each sample, calculate the confidence intervals: 99% for gh, 99% for sector sums, 99% and 97% for sector up-down differences
   c. compute means of limits from the 500 samples
---------------------------------------------------------------------------------------------------------------------------------------
Apply vf criteria to detect glaucoma defect
}
\description{
Analyse a visual field for the presence of a defect based on
 the following criteria: Hodapp-Parish-Anderson 2 (HAP2), United Kington
 Glaucoma Treatment Study (UKGTS), Glaucoma Hemifield Test (GHT), Foster,
 and/or Low-pressure Glaucoma Treatment Study (LoGTS)
 
 All the visual field defect criteria can be run individually with the
 following functions: (HERMAN: this is suggested text. Change at will)
\itemize{
  \item\code{ght} HERMAN: GIVE DETAILS
  \item\code{hap2} HERMAN: GIVE DETAILS
  \item\code{ukgts} HERMAN: GIVE DETAILS
  \item\code{foster} HERMAN: GIVE DETAILS
  \item\code{logts} HERMAN: GIVE DETAILS
}
}
\examples{
# (HERMAN: you need to add some workable examples here of all the
# functions. Check file vf.R or vfplot.R for examples on how)
}
