% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfloaders.R
\name{loadhfaxml}
\alias{loadhfaxml}
\alias{loadhfadicom}
\alias{loadoctopus}
\alias{loadhfaxmlbatch}
\alias{loadhfadicombatch}
\title{Loaders from perimeters}
\usage{
loadhfaxml(file, type = "pwg", repeated = mean)

loadhfadicom(file, type = "pwg", repeated = mean)

loadoctopus(file, type = "pwg", repeated = mean, dateFormat = "\%d.\%m.\%Y")

loadhfaxmlbatch(file, repeated = mean)

loadhfadicombatch(file, repeated = mean)
}
\arguments{
\item{file}{file to load (it is a XML extracted with the export license for
the HFA loader), a CSV outputed by the Eyesuite software for the Octopus
perimeter and a CSV with two columns for the batch HFA loader. The two columns
for the batch HFA loader must be named `\code{file}` and `\code{type}` and must
have the full file name (file path + name) of each XML file to be loaded and the
corresponding patient type, respectively}

\item{type}{type of patient. It can be `\code{ctr}` (for control or healthy
subject-eye) or `\code{pwg}` (for patient with glaucoma) or other}

\item{repeated}{function to apply if there are repeated values in a particular location}

\item{dateFormat}{format to be used for date. Its default value
is \code{\%d.\%m.\%Y}}
}
\value{
visual field data
}
\description{
Functions to load from commercial perimeters
}
\details{
The XML loader for the Humphrery Field Analyser (HFA) by Carl Zeiss Meditec
is essentially a XML parser that reads in the XML generated with the scientific
export license. The DICOMM loader is also a parser to read HFA data generated in a
DICOMM file. The loader for the Octopus perimeter by Haag-Streit is a csv reader
from files generated with the Eyesuite software
}
