% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfpwgRetest24d2.R
\docType{data}
\name{vfpwgRetest24d2}
\alias{vfpwgRetest24d2}
\title{Short-term retest static automated perimetry data}
\format{
See section \code{Structure of visual fields data} in \code{\link{vfdesc}}
}
\usage{
vfpwgRetest24d2
}
\description{
Thirty patients recruited from the glaucoma clinics at the
Queen Elizabeth Health Sciences Centre in Halifax, Nova Scotia. Each
patient underwent 12 visual fields in 12 consecutive weekly sessions.
}
\references{
P. H. Artes, N. O'Leary, M. T. Nicolela, B. C. Chauhan, and D. P. Crabb.
\emph{Visual field progression in glaucoma: What is the specificity of the
guided progression analysis?} American Academy of Ophthalmology,
121(10):2023-2027, 2014.
}
\seealso{
\code{\link{vfpwgSunyiu24d2}}, \code{\link{vfctrIowaPC26}},
         \code{\link{vfctrIowaPeri}}, \code{\link{vfctrSunyiu10d2}},
         \code{\link{vfctrSunyiu24d2}}
}
\keyword{dataset}
