% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfpwgSunyiu24d2.R
\docType{data}
\name{vfpwgSunyiu24d2}
\alias{vfpwgSunyiu24d2}
\title{Series of 24-2 static automated perimetry data for a patient with glaucoma}
\format{
See section \code{Structure of visual fields data} in \code{\link{vfdesc}}
}
\usage{
vfpwgSunyiu24d2
}
\description{
This is real data for the right and left eyes, but the age has been changed
to protect anonymity of the subject. Courtesy of William H Swanson and Mitch W Dul
}
\seealso{
\code{\link{vfctrIowaPC26}}, \code{\link{vfctrIowaPeri}},
         \code{\link{vfctrSunyiu10d2}}, \code{\link{vfctrSunyiu24d2}},
         \code{\link{vfpwgRetest24d2}}
}
\keyword{dataset}
