\name{vfShafi2011}
\alias{vfShafi2011}
\docType{data}
\title{SUNY-IU control data for static automated perimetry 10-2 SITA Standard}
\description{
SUNY-IU control data for static automated perimetry 10-2 SITA Standard
}
\usage{data( vfShafi2011 )}
\format{It is a \code{\link{vf}}-object}

\references{
[1] H. J. Wyatt, M. W. Dul, and W. H. Swanson. \emph{Variability of visual field measurements is correlated with the gradient of visual sensitivity}. Vision Research, 47, 2007.

[2] A. Shafi, W. H. Swanson, and M. W. Dul. \emph{Structure and Function in Patients with Glaucomatous Defects Near Fixation}. Optometry and Vision Science, 88, 2011.
}

\keyword{datasets}

\author{Ivan Marin-Franch, William H Swanson, Harry J Wyatt, Mitchell W Dul}