\name{visualFields-package}
\alias{visualFields}
\title{Statistical methods for visual fields}
\docType{package}
\encoding{UTF-8}
\description{
This is a collection of tools for analyzing the field of vision. It provides a framework for the development and use of innovative methods for visualization, statistical analysis, and clinical interpretation of visual-field loss and its change over time. \code{visualFields} is intended to be a tool for collaborative research
}

\seealso{\code{OPI}, the Open Perimetry Initiative \url{http://people.eng.unimelb.edu.au/aturpin/opi/index.html}}

\references{
[1] Marín-Franch I & Swanson WH. The visualFields package: A tool for analysis and visualization of visual fields. Journal of Vision, 2013, 13(4):10, 1-12

[2] Turpin A, Artes PH, & McKendrick AM. The Open Perimetry Interface: An enabling tool for clinical visual psychophysics. Journal of Vision, 2012, 12(11):22, 21–25

This work was supported by NIH grant number \bold{R01EY007716}
}

\author{Ivan Marin-Franch}