% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waller}
\alias{waller}
\title{Rabies month of first appearance across Connecticut.}
\format{
A data frame with 168 rows and 3 variables:
\describe{
  \item{x}{Longitudinal distance (in km) from first township of appearance}
  \item{y}{Latitudinal distance (in km) from first township of appearance}
  \item{month}{month since first appearance in the state in March 1991}
}
}
\source{
Waller, L. A., & Gotway, C. A. (2004). Applied spatial statistics for public health data (Vol. 368). John Wiley & Sons.
}
\usage{
waller
}
\description{
First month of report of raccoon rabies for 168 townships in Connecticut
from March 1991 through January 1995.
}
\keyword{datasets}
