% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{wattsStrogatz}
\alias{wattsStrogatz}
\title{Function to generate a Watts-Strogatz network}
\usage{
wattsStrogatz(N, K, Prw)
}
\arguments{
\item{N}{the number of nodes}

\item{K}{the number of neighbors to which each node is connected so degree = 2*K}

\item{Prw}{the rewiring probability}
}
\value{
An object of class CM (contact matrix)
}
\description{
Function to generate a Watts-Strogatz network
}
\examples{
cm2=wattsStrogatz(N=20, K=4, Prw=.3)
}
