% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_pipe.R
\name{whole_pipe}
\alias{whole_pipe}
\title{Some title}
\usage{
whole_pipe(
  path_total,
  filename_chem = "Water_Chemistry_R_Package.csv",
  filename_poll = "Pollution_Site_R_Package.csv",
  upstream_thresh = 5,
  downstream_lower_thresh = 0,
  downstream_upper_thresh = 50,
  num_cores = 1,
  permission = "no"
)
}
\arguments{
\item{path_total}{description of this parameter}

\item{filename_chem}{description of this parameter}

\item{filename_poll}{description of this parameter}

\item{upstream_thresh}{description of this parameter}

\item{downstream_lower_thresh}{description of this parameter}

\item{downstream_upper_thresh}{description of this parameter}

\item{num_cores}{description of this parameter}

\item{permission}{description of this parameter}
}
\value{
description of the return value
}
\description{
A description of this function with inline latex \eqn{2^3=8}
test for displayed equations \deqn{\frac{1}{2}\sum_{i=1}^n\log(x+y)}.
We can also use links, tables, lists and Character formattings such as bold and italic.
}
