% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
wiki_graph
}
\description{
wiki_graph data is for testing dijstra method.
}
\examples{
data(wiki_graph)
v1 <- wiki_graph$v1
v2 <- wiki_graph$v2
w <- wiki_graph$w
}
\references{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{wikipedia page}
}
\keyword{datasets}
