% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Worldwide_Population.R
\docType{class}
\name{Worldwide_Pollution-class}
\alias{Worldwide_Pollution-class}
\alias{Worldwide_Pollution}
\title{Title Worldwide_Pollution}
\description{
This class creates one object which can give access in an API.The user can choose between four countries(Greece,Italy,Sweden,Turkey),
and manipulate data for air pollution such as value pm5.
}
\section{Fields}{

\describe{
\item{\code{responses}}{list. The list of responses comming from API}

\item{\code{root_url}}{character. The base url of the API}

\item{\code{countries}}{list. A list of the user's input}
}}


