\name{xmlitem}
\alias{xmlitem}
\title{extracts from a loaded XML file the info of a tag}
\description{extracts from a loaded XML file the info of a tag}
\usage{xmlitem( tag, xmllines, capitalize = TRUE )}
\arguments{
  \item{tag}{tag to look at}
  \item{xmllines}{lines from loaded XML files}
  \item{capitalize}{Whether we need to capitalize or not. Default is \code{TRUE}}
}
\details{extracts from a loaded XML file the info of a tag}
\seealso{\code{\link{loadvfxml}}, \code{\link{loadvfcsv}},\code{\link{xmlblock}}}

\author{Ivan Marin-Franch}