% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_clean.R
\name{zip_clean}
\alias{zip_clean}
\title{Function cleans up zipcodes that contain extra information.}
\usage{
zip_clean(Zip = ..., blind = FALSE, blind_chr = "xx")
}
\arguments{
\item{blind}{currently non functional but returns logical var to
tell function to blind}

\item{blind_chr}{the characters exactly n and type to replace the
last n digits}

\item{zip}{a single of a vector of zipcodes with junk in fields.
Very basic}
}
\value{
cleaned zips
}
\description{
Given a zip code or a vector of zipcodes, the function extracts the
zip code and then returns a clean zip code.  Blinding will remove
the last n char as determined by blind_chr
}
\examples{
\dontrun{
zip_test  <- "37692 x"

zip_clean(zip_test) # Basic function

}

}
