\name{AGEP-package}
\Rdversion{1.1}
\alias{AGEP-package}
\alias{AGEP}
\docType{package}
\title{
Allows alignment of single profile against a reference data
}
\description{
Allows transformation of two-dimensional data matrix to a reference densities and alignment of single profile (vector) against it.
}
\details{
\tabular{ll}{
Package: \tab AGEP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-06-01\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Sami Kilpinen, Kalle Ojala

Maintainer:
sami.k.kilpinen@helsinki.fi
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
