
\name{data}
\alias{data}

%\docType{package}
\title{
  Data for the \code{\link{MutScot}} example
}
\description{
 Data for the \code{\link{MutScot}} example.
}
\details{ 
  The object \code{data} is a data frame with columns for the sample id, 
   chromosome, position (hg19), reference allele, and alternate allele.
}
\examples{
  data(data, package="MUTSCOT")
  data[1:5, ]
}
\keyword{data}
%\seealso{ \code{\link{subhmm}} }



