% --- Source file: data.Rd ---
\name{data}
\alias{data}
\alias{V}
\alias{L}
\alias{W}
\alias{H}

\title{
    Data for examples
}
\description{
    Example input data objects
}
\details{ 
  Contains example \code{L}, \code{V}, \code{W}, and \code{H} matrices
  for examples.
}
%\examples{}
\keyword{data}
