
\name{data}
\alias{data}
\alias{logR}
\alias{logOR}


%\docType{package}
\title{
  Data for examples
}
\description{
 Data for examples.
}
\details{ The object contains \code{logR} and \code{logOR} values
  for \code{\link{subhmm}}.
}
\examples{

 data(data, package="subHMM")

 # Display some of the data
 logR[1:10]
 logOR[1:10]
}
\keyword{data}
\seealso{ \code{\link{subhmm}} }



