% --- Source file: data.Rd ---
\name{data}
\alias{data}
\alias{results}

\title{
  Data for examples
}
\description{
 Example input data and results
}
\details{ Contains an example input data object and an example 
  matrix of results.
}
\examples{

 data(data, package="SUITOR")

 # Display a subset of data objects
 data[1:5, 1:5]
 results[1:10, ]

}
\keyword{data}
\seealso{ \code{\link{suitor}} }



