% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.behavioural.base}
\alias{organise.behavioural.base}
\title{Save RT and Accuracy split by specified columns.}
\usage{
organise.behavioural.base(fixreport_df, grouping_column_list,
  response_period_start = "", prefix_label = "")
}
\arguments{
\item{fixreport_df}{Fixation report.}

\item{grouping_column_list}{List of columns to split by.}

\item{response_period_start}{Message that starts the RT timer.}

\item{prefix_label}{Prefix output with a label.}
}
\value{
Summarised behavioural information as a data.table. This is also saved to disk.
}
\description{
Save RT and Accuracy split by specified columns.
}
\examples{
data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
                               message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
                               message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_RESPONSE")

# NOW MARK UP FIXATION CONTINGENCIES
foo<-organise.message.fix_contingencies(foo, list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))

# RANDOM TRIAL TO CHECK THINGS OUT
organise.checks.random_trial(foo)

# FIX CONTINGENCIES
organise.contingencies.descriptives(foo)

# REMOVE MISSING EVENTS - HERE, TRIALS WHICH LACKED A RESPOSNE
foo <- organise.message.removals(fixreport_df=foo, 
                                 required_message_list=list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))
# REMOVES NO TRIALS = EXCELLENT!

organise.contingencies.descriptives(foo) # THIS SHOWS WE HAVE NO UNCLASSIFIED FIXATIONS, GOOD!

# BY-TRIAL BEHAVIOURAL DATA
organise.behavioural.base(fixreport_df = foo, list( 'TRIALTYPE_TEXT'), 
                          response_period_start="SYNCTIME", 
                          prefix_label="BYTRIAL")
}
