% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.checks.random_trial}
\alias{organise.checks.random_trial}
\title{Display a randomly selected trial for detailed checks.}
\usage{
organise.checks.random_trial(fixreport_df)
}
\arguments{
\item{fixreport_df}{object Input fixation report.}
}
\value{
Single trial as a data.table, plus the same trial printed to the console.
}
\description{
Display a randomly selected trial for detailed checks.
}
\examples{
data(fixationreport)
organise.checks.random_trial(fixationreport)
}
