% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.message.fix_contingencies}
\alias{organise.message.fix_contingencies}
\title{Oganise and markup fixation contingencies.}
\usage{
organise.message.fix_contingencies(fixreport_df, ordered_message_list)
}
\arguments{
\item{fixreport_df}{Fixation report.}

\item{ordered_message_list}{List of messages to markup, in temporal order at which they occurred.}
}
\value{
Marked-up fixation report data.table.
}
\description{
Oganise and markup fixation contingencies.
}
\examples{

data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
   message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
   message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_RESPONSE")

# NOW MARK UP FIXATION CONTINGENCIES
foo<-organise.message.fix_contingencies(foo, list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))
}
