% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.message.removals}
\alias{organise.message.removals}
\title{Remove trials which fail to have all of the listed messages.}
\usage{
organise.message.removals(fixreport_df, required_message_list)
}
\arguments{
\item{fixreport_df}{Fixation report.}

\item{required_message_list}{List of messages required for each trial.}
}
\value{
Updated fixation report with trials removed as a data.table.
}
\description{
Remove trials which fail to have all of the listed messages.
}
\examples{
data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
                               message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
                               message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_RESPONSE")

# NOW MARK UP FIXATION CONTINGENCIES
foo<-organise.message.fix_contingencies(foo, list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))

# RANDOM TRIAL TO CHECK THINGS OUT
organise.checks.random_trial(foo)

# FIX CONTINGENCIES
organise.contingencies.descriptives(foo)

# REMOVE MISSING EVENTS - HERE, TRIALS WHICH LACKED A RESPOSNE
foo <- organise.message.removals(fixreport_df=foo, 
                                 required_message_list=list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))
# REMOVES NO TRIALS = EXCELLENT!

organise.contingencies.descriptives(foo) 
# THIS LISTS NO UNCLASSIFIED FIXATIONS WHICH IS GOOD BECAUSE IT MEANS WE ACCOUNT FOR 
# EVERYTHING THAT TOOK PLACE IN OUR TRIALS.
}
