% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{plot.netSir}
\alias{plot.netSir}
\title{Function to plot a netSir object}
\usage{
\method{plot}{netSir}(x, ...)
}
\arguments{
\item{x}{an object of class netSir}

\item{...}{other arguments}
}
\description{
Function to plot a netSir object
}
\seealso{
\code{\link{netSir}}
}
