% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gams.R
\name{run_gams}
\alias{run_gams}
\title{Auxiliary functions for executing GAMS}
\usage{
run_gams(model_file, args, gams_args = "LogOption=4")
}
\arguments{
\item{args}{additional arguments for the CLI call to gams}

\item{model}{model name}

\item{ingdx}{input gdx file}

\item{outgdx}{output gdx file}

\item{model_pth}{the path to the gams file}
}
\description{
Auxiliary functions for executing GAMS
}
