% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfreports.R
\name{vfsfa}
\alias{vfsfa}
\alias{vfsfashiny}
\title{Single Field Reporting}
\usage{
vfsfa(vf, file, ...)

vfsfashiny(vf, ...)
}
\arguments{
\item{vf}{visual field data}

\item{file}{The pdf file name where to save the one-page reports of single field analysis}

\item{...}{other graphical arguments}
}
\value{
No return value
}
\description{
Generates of one-page reports of single field analyses
}
\details{
\itemize{
  \item\code{vfsfa} saves a pdf with one-page reports of single field analyses
  \item\code{vfsfashiny} generates interactive one-page reports of single field
    analyses based on Shiny
}
}
