% --- Source file: SimData.Rd ---
\name{SimData}
\alias{SimData}

\title{
    Data for examples
}
\description{ 
  Simulated data for examples
}
\details{
  This file consists of the list \code{SimData} with the following objects: 
  \itemize{
    \item{V} {: Simulated mutation catalog matrix of size 96 by 10027}
    \item{L} {: Panel size matrix of size 96 by 10027}
    \item{W_TMB} {: Tumor mutation burden based signatures size 96 by 76}
    \item{TrueH} {: Simulated signature activity matrix of size 6 by 10027 used to generate V}
    \item{TrueW_TMB} {: Tumor mutation burden based signatures size 96 by 6 used to generate V}
  }
 }
%\examples{}
\keyword{data}
