% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfevent.R
\name{vfdefect}
\alias{vfdefect}
\title{Event criteria for detecting glaucomatous damage}
\usage{
vfdefect(
  vf,
  criteria = "all",
  ghtlims = NA,
  vfctr = visualFields::vfctrSunyiu24d2,
  tpattern = "p24d2"
)
}
\arguments{
\item{vf}{Visual fields to be analyzed as a standard vfobject}

\item{criteria}{select to use \code{all} (default) or one criteria:
\code{ght}, \code{hap2}, \code{foster}, \code{ukgts}, or \code{logts}}

\item{ghtlims}{a list containing limits of normality for the Glaucoma Hemifield 
Test (GHT); this list is returned by \code{vfdefect} function}

\item{vfctr}{control data, which is a vf object containing normative vfs}

\item{tpattern}{pattern of visual field test, either 24-2 or 30-2}
}
\value{
A list of 2 objects: 
\itemize{
 \item \code{$ght.limits} limits of normality required by the Glaucoma Hemifield Test (GHT) 
 \item \code{$results} results of detected glaucoma damage in \code{vf} parameter using the 
 user-selected \code{criteria} 
}
}
\description{
Analyze a visual field for the presence of a defect based on
 the following criteria: Glaucoma Hemifield Test (GHT), Hodapp-Anderson-Parrish 2 (HAP2), 
 Foster, United Kingdom, Glaucoma Treatment Study (UKGTS), 
 and/or Low-pressure Glaucoma Treatment Study (LoGTS)
}
\details{
The \code{vfdefect} function is a wrapper that contains all 5 event criteria. The user 
 can select which \code{criteria} to apply to the \code{vf} by passing \code{all} (default), 
 \code{ght}, \code{hap2}, \code{foster}, \code{ukgts}, or \code{logts}.
 
 The user can provide their own limits of normality \code{ghtlims} for GHT, or let the 
 function compute the limits by providing a vf object \code{vfctr} that contains normal visual fields.

Event criteria description 
\itemize{
  \item GHT: Decision tree focused on the comparison of 5 regions in the superior hemifield 
  with their mirror images in the inferior hemifield. Criteria commonly considered positive 
  if the output is "outside normal limits" [1]
  \item HAP2: GHT "Outside normal limits" \emph{OR} a cluster of 3 points depressed at P<0.05 level, 
  one of which at P<0.01 on the pattern deviation plot \emph{OR} PSD at P<5% [2]
  \item Foster: GHT "outside normal limits" \emph{AND} a cluster of 3 contiguous points depressed at 
  P<0.05 level on the pattern deviation plot [3]
  \item UKGTS: a cluster of 2 points depressed at P<0.01 level \emph{OR} a cluster of 3 points 
  depressed at P<0.05 level \emph{OR} a 10 dB difference across the nasal horizontal midline 
  at 2 or more adjacent points on the total deviation plot [4]
  \item LoGTS: a cluster of 3 points depressed at < -8 dB \emph{OR} a cluster of 2 
  points depressed at < -10 dB on the total deviation plot [5]
}
}
\examples{
 # check visual fields for defect using all 5 criteria
 result <- vfdefect( vfpwgSunyiu24d2 )
 
 # check visual field for defect using the GHT criteria
 result <- vfdefect( vfpwgSunyiu24d2, criteria = "ght")
 
 # speed up computation when re-using the same ght limits of normality
 result1 <- vfdefect( vfpwgSunyiu24d2 )
 result2 <- vfdefect( vfpwgRetest24d2, ghtlims = result1$ght.limits )
 
 # use a custom normative dataset with the criteria to detect glaucoma defect:
 # 1.prepare the statistical environment with your normative dataset
 # 2.check visual fields for defect using all 5 criteria
 setnv( nvgenerate( vfctrSunyiu24d2 ) )
 result <- vfdefect( vfpwgSunyiu24d2, vfctr = vfctrSunyiu24d2 )
}
\references{
[1] Asman P, Heijl A. \emph{Glaucoma Hemifield Test: Automated Visual Field Evaluation.} 
 Arch Ophthalmol. 1992;110(6):812-819.
 
 [2] Hodapp E, Parrish RK, Anderson D. \emph{Clinical Decisions in Glaucoma.} 
 1st ed. Mosby, Inc.; 1993.
 
 [3] Foster PJ, Buhrmann R, Quigley HA, Johnson GJ. \emph{The definition and 
 classification of glaucoma in prevalence surveys.} Br J Ophthalmol. 2002;86(2):238-42.
 
 [4] Garway-Heath DF, Lascaratos G, Bunce C, Crabb DP, Russell RA, Shah A. 
 \emph{The United Kingdom glaucoma treatment study: A multicenter, randomized, 
 placebo-controlled clinical trial: Design and methodology.} Ophthalmology. 
 2013;120(1):68-76.
 
 [5] Krupin T, Liebmann JM, Greenfield DS, Rosenberg LF, Ritch R, Yang JW. \emph{The 
 Low-pressure Glaucoma Treatment Study (LoGTS): Study design and baseline characteristics 
 of enrolled patients.} Ophthalmology. 2005;112(3):376-385.
}
\author{
Herman Stubeda, Dalhousie University Medical School
}
